### 2025-09-06 d30272a
Update paper.md
