### 2025-09-08 d20905c
Update paper.md
