### 2025-09-08 bd5c567
Update paper.bib
