+++ paper/paper.bib
@@ -57,3 +57,14 @@ publisher={Elsevier},
 doi={10.1016/j.envsoft.2017.09.020}
 }
 
+@article{kamath2023human,
+title={Human heat health index (H3I) for holistic assessment of heat hazard and mitigation strategies beyond urban heat islands},
+author={Kamath, Harsh G and Martilli, Alberto and Singh, Manmeet and Brooks, Trevor and Lanza, Kevin and Bixler, R Patrick and Coudert, Marc and Yang, Zong-Liang and Niyogi, Dev},
+journal={Urban Climate},
+volume={52},
+pages={101675},
+year={2023},
+publisher={Elsevier}
+doi={10.1016/j.uclim.2023.101675}
+}
+
