### 2025-09-08 cb3177d
Update paper.md
