+++ paper/paper.md
@@ -47,7 +47,7 @@ We present Solar and LongWave Environmental Irradiance Geometry-Graphics Process
 
 # Statement of need
 
-The original SOLWEIG model [@lindberg2008solweig] was developed to calculate T<sub>MRT</sub> over small geographical areas in cities. At this spatial scale, based on the time required for computation, the model could be run on CPUs. However, for city-scale thermal comfort estimation, the model can be accelerated using a GPU. Specifically, the calculation of the SVF (which is the most time-consuming step), T<sub>MRT</sub> and UTCI can be processed on a GPU. Currently, there is a tool that computes SVF on GPU but it uses Python to read the inputs and write the output rasters, while the SVF calculation on GPU is done by interfacing with a C-language code [@li2021gpu]. Thus, there was a need for a Python-based end-to-end framework to run SOLWEIG on a GPU. Our framework is implemented in PyTorch, which enables automatic selection of the GPU when available. 
+The original SOLWEIG model [@lindberg2008solweig] was developed to calculate T<sub>MRT</sub> over small geographical areas in cities. At this spatial scale, based on the time required for computation, the model could be run on CPUs [@kamath2023human]. However, for city-scale thermal comfort estimation, the model can be accelerated using a GPU. Specifically, the calculation of the SVF (which is the most time-consuming step), T<sub>MRT</sub> and UTCI can be processed on a GPU. Currently, there is a tool that computes SVF on GPU but it uses Python to read the inputs and write the output rasters, while the SVF calculation on GPU is done by interfacing with a C-language code [@li2021gpu]. Thus, there was a need for a Python-based end-to-end framework to run SOLWEIG on a GPU. Our framework is implemented in PyTorch, which enables automatic selection of the GPU when available. 
 
 
 # Functionality 
