### 2025-09-08 2f21051
Update paper.bib
