+++ paper/paper.bib
@@ -64,7 +64,7 @@ journal={Urban Climate},
 volume={52},
 pages={101675},
 year={2023},
-publisher={Elsevier}
+publisher={Elsevier},
 doi={10.1016/j.uclim.2023.101675}
 }
 
