+++ solweig_gpu/Tgmaps_v1.py
@@ -1,3 +1,15 @@
+#SOLWEIG-GPU: GPU-accelerated SOLWEIG model for urban thermal comfort simulation
+#Copyright (C) 2022–2025 Harsh Kamath and Naveen Sudharsan
+
+#This program is free software: you can redistribute it and/or modify
+#it under the terms of the GNU General Public License as published by
+#the Free Software Foundation, either version 3 of the License, or
+#(at your option) any later version.
+
+#This program is distributed in the hope that it will be useful,
+#but WITHOUT ANY WARRANTY; without even the implied warranty of
+#MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+#GNU General Public License for more details.
 import numpy as np
 
 def Tgmaps_v1(lc_grid, lc_class):
@@ -24,5 +36,6 @@ def Tgmaps_v1(lc_grid, lc_class):
     TgK_wall = lc_class[wall_pos, 3]
     Tstart_wall = lc_class[wall_pos, 4]
     TmaxLST_wall = lc_class[wall_pos, 5]
+
 
     return TgK, Tstart, alb_grid, emis_grid, TgK_wall, Tstart_wall, TmaxLST, TmaxLST_wall
