### 2025-09-15 0a0e661
Update calculate_utci.py
