+++ solweig_gpu/calculate_utci.py
@@ -1,3 +1,16 @@
+#SOLWEIG-GPU: GPU-accelerated SOLWEIG model for urban thermal comfort simulation
+#Copyright (C) 2022–2025 Harsh Kamath and Naveen Sudharsan
+
+#This program is free software: you can redistribute it and/or modify
+#it under the terms of the GNU General Public License as published by
+#the Free Software Foundation, either version 3 of the License, or
+#(at your option) any later version.
+
+#This program is distributed in the hope that it will be useful,
+#but WITHOUT ANY WARRANTY; without even the implied warranty of
+#MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+#GNU General Public License for more details.
+
 import torch
 
 # Function to computed UTCI using 6th order polynomial approximation. 
@@ -255,4 +268,5 @@ def utci_calculator(Ta, RH, Tmrt, va10m):
     UTCI_approx[valid_mask] = utci_polynomial(D_Tmrt, Ta[valid_mask], va, Pa)
 
     return UTCI_approx
+
 
