### 2025-09-19 944336d
Update shadow.py
add license
