### 2025-09-19 8c2cf0e
Update solweig_gpu.py
add license
