### 2025-09-19 e88cbe7
Update utci_process.py
Add license
