### 2025-09-19 3dad830
Update walls_aspect.py
Add license
