### 2025-09-22 e79004b
Update calculate_utci.py
add license
