### 2025-09-22 8ae757e
Update solweig_gpu_gui.py
