### 2025-09-30 #8944 @kthyng
@nvnsudharsan Yes both are required — a readme is not sufficient, and you'll need detailed docstrings too. For testing, both unit tests and examples demonstrating how to use the code would be best. A nice way to make your software more usable is with hosted docs like on readthedocs — you can compile your docs and your notebook examples all together that way.
