### 2025-10-07 fb69dab
add docs and examples
