+++ tests/test_walls_aspect.py
@@ -1,0 +1,86 @@
+"""
+Test suite for the walls_aspect module.
+
+This module tests the wall height and aspect calculation functionality.
+"""
+
+import unittest
+import numpy as np
+import os
+import tempfile
+from osgeo import gdal, osr
+
+
+class TestWallCalculation(unittest.TestCase):
+    """Test wall height calculation."""
+
+    def setUp(self):
+        """Set up test fixtures."""
+        self.temp_dir = tempfile.mkdtemp()
+
+    def tearDown(self):
+        """Clean up test fixtures."""
+        # Clean up temporary files
+        pass
+
+    def create_simple_building_dsm(self):
+        """Create a simple building DSM for testing."""
+        # Create a 10x10 raster with a simple building
+        data = np.zeros((10, 10), dtype=np.float32)
+        # Add a 5m tall building in the center
+        data[4:7, 4:7] = 5.0
+        
+        return data
+
+    def test_wall_height_simple_building(self):
+        """Test wall height calculation for a simple building."""
+        dsm = self.create_simple_building_dsm()
+        
+        # Expected: walls should be detected at the edges of the building
+        # with height of 5m
+        # This is a placeholder - actual test would call the wall calculation function
+        pass
+
+    def test_wall_height_no_buildings(self):
+        """Test wall height calculation with no buildings."""
+        dsm = np.zeros((10, 10), dtype=np.float32)
+        
+        # Expected: no walls should be detected
+        pass
+
+    def test_wall_height_multiple_buildings(self):
+        """Test wall height calculation with multiple buildings."""
+        # Create DSM with multiple buildings of different heights
+        pass
+
+
+class TestAspectCalculation(unittest.TestCase):
+    """Test aspect (orientation) calculation."""
+
+    def test_aspect_north_facing(self):
+        """Test aspect calculation for north-facing wall."""
+        # Expected aspect should be close to 0° (north)
+        pass
+
+    def test_aspect_east_facing(self):
+        """Test aspect calculation for east-facing wall."""
+        # Expected aspect should be close to 90° (east)
+        pass
+
+    def test_aspect_south_facing(self):
+        """Test aspect calculation for south-facing wall."""
+        # Expected aspect should be close to 180° (south)
+        pass
+
+    def test_aspect_west_facing(self):
+        """Test aspect calculation for west-facing wall."""
+        # Expected aspect should be close to 270° (west)
+        pass
+
+    def test_aspect_range(self):
+        """Test that aspect values are within valid range [0, 360]."""
+        pass
+
+
+if __name__ == '__main__':
+    unittest.main()
