### 2025-10-07 1a9579f
add docs and examples
