+++ tests/test_calculate_utci.py
@@ -1,0 +1,73 @@
+"""
+Test suite for the calculate_utci module.
+
+This module tests the UTCI (Universal Thermal Climate Index) calculation.
+"""
+
+import unittest
+import numpy as np
+
+
+class TestUTCICalculation(unittest.TestCase):
+    """Test UTCI calculation functionality."""
+
+    def test_utci_comfortable_conditions(self):
+        """Test UTCI calculation for comfortable conditions."""
+        # This is a placeholder test
+        # Actual implementation would test the utci_calculator function
+        # with known reference values
+        
+        # Example comfortable conditions:
+        # Ta = 25°C, Tmrt = 25°C, wind = 1 m/s, RH = 50%
+        # Expected UTCI should be close to 25°C
+        pass
+
+    def test_utci_hot_conditions(self):
+        """Test UTCI calculation for hot conditions."""
+        # Test with high temperature and high radiation
+        # Expected UTCI should be significantly higher than air temperature
+        pass
+
+    def test_utci_cold_conditions(self):
+        """Test UTCI calculation for cold conditions."""
+        # Test with low temperature and high wind
+        # Expected UTCI should be lower than air temperature (wind chill)
+        pass
+
+    def test_utci_extreme_heat(self):
+        """Test UTCI calculation for extreme heat conditions."""
+        # Test boundary conditions for extreme heat
+        pass
+
+    def test_utci_extreme_cold(self):
+        """Test UTCI calculation for extreme cold conditions."""
+        # Test boundary conditions for extreme cold
+        pass
+
+    def test_utci_input_validation(self):
+        """Test that invalid inputs are handled correctly."""
+        # Test that the function handles out-of-range inputs appropriately
+        pass
+
+
+class TestUTCIArrayProcessing(unittest.TestCase):
+    """Test UTCI calculation with array inputs."""
+
+    def test_utci_2d_array(self):
+        """Test UTCI calculation with 2D arrays."""
+        # Test that the function can process 2D spatial arrays
+        pass
+
+    def test_utci_3d_array(self):
+        """Test UTCI calculation with 3D arrays (spatial + temporal)."""
+        # Test that the function can process 3D arrays
+        pass
+
+    def test_utci_nan_handling(self):
+        """Test that NaN values are handled correctly."""
+        # Test that NaN values in input don't cause errors
+        pass
+
+
+if __name__ == '__main__':
+    unittest.main()
