+++ tests/test_sun_position.py
@@ -1,0 +1,99 @@
+"""
+Test suite for the sun_position module.
+
+This module tests the solar position calculations to ensure accuracy.
+"""
+
+import unittest
+import numpy as np
+import datetime
+
+
+class TestSolarPosition(unittest.TestCase):
+    """Test solar position calculations."""
+
+    def test_solar_position_noon(self):
+        """Test solar position calculation at solar noon."""
+        # This is a placeholder test - actual implementation would require
+        # importing and testing the Solweig_2015a_metdata_noload function
+        # with known reference values
+        pass
+
+    def test_solar_position_sunrise(self):
+        """Test solar position calculation at sunrise."""
+        # Placeholder for sunrise test
+        pass
+
+    def test_solar_position_sunset(self):
+        """Test solar position calculation at sunset."""
+        # Placeholder for sunset test
+        pass
+
+    def test_solar_azimuth_range(self):
+        """Test that solar azimuth is within valid range [0, 360]."""
+        # Placeholder - would test that azimuth values are always in valid range
+        pass
+
+    def test_solar_altitude_range(self):
+        """Test that solar altitude is within valid range [-90, 90]."""
+        # Placeholder - would test that altitude values are always in valid range
+        pass
+
+
+class TestDayLength(unittest.TestCase):
+    """Test day length calculations."""
+
+    def test_summer_solstice(self):
+        """Test day length calculation at summer solstice."""
+        from solweig_gpu.solweig import daylen
+        import torch
+        
+        # Summer solstice (approximately day 172)
+        DOY = torch.tensor(172.0)
+        # Austin, TX latitude
+        XLAT = torch.tensor(30.27)
+        
+        DAYL, DEC, SNDN, SNUP = daylen(DOY, XLAT)
+        
+        # Day length should be greater than 12 hours in northern hemisphere summer
+        self.assertGreater(DAYL.item(), 12.0)
+        # Declination should be positive (northern hemisphere summer)
+        self.assertGreater(DEC.item(), 0.0)
+
+    def test_winter_solstice(self):
+        """Test day length calculation at winter solstice."""
+        from solweig_gpu.solweig import daylen
+        import torch
+        
+        # Winter solstice (approximately day 355)
+        DOY = torch.tensor(355.0)
+        # Austin, TX latitude
+        XLAT = torch.tensor(30.27)
+        
+        DAYL, DEC, SNDN, SNUP = daylen(DOY, XLAT)
+        
+        # Day length should be less than 12 hours in northern hemisphere winter
+        self.assertLess(DAYL.item(), 12.0)
+        # Declination should be negative (northern hemisphere winter)
+        self.assertLess(DEC.item(), 0.0)
+
+    def test_equinox(self):
+        """Test day length calculation at equinox."""
+        from solweig_gpu.solweig import daylen
+        import torch
+        
+        # Spring equinox (approximately day 80)
+        DOY = torch.tensor(80.0)
+        # Austin, TX latitude
+        XLAT = torch.tensor(30.27)
+        
+        DAYL, DEC, SNDN, SNUP = daylen(DOY, XLAT)
+        
+        # Day length should be approximately 12 hours at equinox
+        self.assertAlmostEqual(DAYL.item(), 12.0, delta=0.5)
+        # Declination should be close to zero
+        self.assertAlmostEqual(DEC.item(), 0.0, delta=5.0)
+
+
+if __name__ == '__main__':
+    unittest.main()
