+++ tests/test_integration.py
@@ -1,0 +1,159 @@
+"""
+Integration tests for SOLWEIG-GPU.
+
+These tests verify that the entire workflow functions correctly from end to end.
+"""
+
+import unittest
+import os
+import tempfile
+import shutil
+import numpy as np
+from osgeo import gdal, osr
+
+
+class TestEndToEndWorkflow(unittest.TestCase):
+    """Test the complete workflow from input to output."""
+
+    def setUp(self):
+        """Set up test fixtures."""
+        self.temp_dir = tempfile.mkdtemp()
+        self.base_path = os.path.join(self.temp_dir, 'test_data')
+        os.makedirs(self.base_path, exist_ok=True)
+
+    def tearDown(self):
+        """Clean up test fixtures."""
+        if os.path.exists(self.temp_dir):
+            shutil.rmtree(self.temp_dir)
+
+    def create_minimal_test_data(self):
+        """Create minimal test data for integration testing."""
+        # Create simple test rasters
+        width, height = 50, 50
+        
+        # Building DSM
+        building_dsm = np.zeros((height, width), dtype=np.float32)
+        building_dsm[20:30, 20:30] = 10.0  # 10m building
+        
+        # DEM (flat terrain)
+        dem = np.zeros((height, width), dtype=np.float32)
+        
+        # Trees (some vegetation)
+        trees = np.zeros((height, width), dtype=np.float32)
+        trees[10:15, 10:15] = 5.0  # 5m trees
+        
+        # Save rasters
+        self._save_raster(building_dsm, 'Building_DSM.tif')
+        self._save_raster(dem, 'DEM.tif')
+        self._save_raster(trees, 'Trees.tif')
+
+    def _save_raster(self, data, filename):
+        """Helper function to save a raster."""
+        driver = gdal.GetDriverByName('GTiff')
+        filepath = os.path.join(self.base_path, filename)
+        dataset = driver.Create(filepath, data.shape[1], data.shape[0], 1, gdal.GDT_Float32)
+        
+        # Set geotransform (1m pixel size)
+        geotransform = (0, 1.0, 0, 0, 0, -1.0)
+        dataset.SetGeoTransform(geotransform)
+        
+        # Set projection (UTM Zone 14N - Austin, TX)
+        srs = osr.SpatialReference()
+        srs.ImportFromEPSG(32614)
+        dataset.SetProjection(srs.ExportToWkt())
+        
+        # Write data
+        band = dataset.GetRasterBand(1)
+        band.WriteArray(data)
+        
+        dataset.FlushCache()
+        dataset = None
+
+    def create_minimal_met_file(self):
+        """Create a minimal meteorological file for testing."""
+        met_file = os.path.join(self.base_path, 'test_met.txt')
+        
+        # Create a simple met file with one day of data
+        header = "Year DOY Hour Min Ta RH G D I Ws Wd P"
+        data_lines = []
+        for hour in range(24):
+            # Simple diurnal cycle
+            ta = 20 + 10 * np.sin((hour - 6) * np.pi / 12)  # Temperature
+            rh = 60  # Relative humidity
+            g = max(0, 800 * np.sin((hour - 6) * np.pi / 12))  # Global radiation
+            d = g * 0.3  # Diffuse radiation
+            i = g - d  # Direct radiation
+            ws = 2.0  # Wind speed
+            wd = 180  # Wind direction
+            p = 1013  # Pressure
+            
+            data_lines.append(f"2020 225 {hour} 0 {ta:.1f} {rh} {g:.1f} {d:.1f} {i:.1f} {ws} {wd} {p}")
+        
+        with open(met_file, 'w') as f:
+            f.write(header + '\n')
+            f.write('\n'.join(data_lines))
+        
[…]

[…]
