### 2025-10-10 69fd208
Update examples.md
