### 2025-10-10 aeffe91
Update installation.md
correction in the package name
