+++ README.md
@@ -76,7 +76,7 @@ UMEP journal reference: Lindberg, F., Grimmond, C.S.B., Gabey, A., Huang, B., Ke
 ```bash
 conda create -n solweig python=3.10
 conda activate solweig
-conda install -c conda-forge gdal cudnn pytorch timezonefinder matplotlib pyqt=5 sip
+conda install -c conda-forge gdal cudnn pytorch timezonefinder matplotlib pyqt5 sip #cudnn is required only if you are using nvidia GPU
 pip install solweig-gpu
 
 ```
