### 2025-10-10 f52b6f5
Update README.md
correction for package installation
