### 2025-10-14 e01c03c
Update examples.md
added ERA5 download script
