### 2025-10-14 94e9e4e
Update examples.md
