### 2025-10-14 7a1a08d
Update preprocessor.py
made the era5 part more flexible
