+++ solweig_gpu/preprocessor.py
@@ -195,42 +195,118 @@ def process_era5_data(start_time, end_time, folder_path, output_file="Outfile.nc
     """
     start_time = datetime.datetime.strptime(start_time, "%Y-%m-%d %H:%M:%S")
     end_time = datetime.datetime.strptime(end_time, "%Y-%m-%d %H:%M:%S")
+    
     def saturation_vapor_pressure(T):
         """
         Calculate saturation vapor pressure (in hPa) given temperature T in Celsius.
         """
         return 6.112 * np.exp((17.67 * T) / (T + 243.5))
     
-    instant_file = os.path.join(folder_path, 'data_stream-oper_stepType-instant.nc')
-    accum_file   = os.path.join(folder_path, 'data_stream-oper_stepType-accum.nc')
+    # Use glob to find all NetCDF files in the folder
+    nc_files = glob.glob(os.path.join(folder_path, "*.nc"))
+    if not nc_files:
+        raise FileNotFoundError(f"No NetCDF files found in {folder_path}")
     
-    ds_instant = xr.open_dataset(instant_file)
-    ds_accum   = xr.open_dataset(accum_file)
+    print(f"Found {len(nc_files)} NetCDF files: {[os.path.basename(f) for f in nc_files]}")
     
-    # Generate the correct time array using the provided start and end times (1-hour frequency)
-    time_array = [start_time + timedelta(hours=i) 
-                  for i in range(int((end_time - start_time).total_seconds() // 3600) + 1)]
+    # Open all NetCDF files using mfdataset
+    try:
+        ds = xr.open_mfdataset(nc_files, combine='by_coords')
+    except Exception as e:
+        print(f"Warning: mfdataset failed with error: {e}")
+        print("Trying to open files individually...")
+        # Fallback: try to open files individually and combine
+        datasets = []
+        for file in nc_files:
+            try:
+                ds_temp = xr.open_dataset(file)
+                datasets.append(ds_temp)
+            except Exception as e2:
+                print(f"Could not open {file}: {e2}")
+                continue
+        
+        if not datasets:
+            raise ValueError("Could not open any NetCDF files")
+        
+        # Combine datasets
+        ds = xr.merge(datasets)
     
-    temperatures = ds_instant['t2m'].values
-    dew_points   = ds_instant['d2m'].values
+    # Identify the time coordinate
+    time_coord_name = None
+    possible_time_names = ['time', 'valid_time', 'Time', 'Valid_time']
     
-    surface_pressures = ds_instant['sp'].values 
+    for time_name in possible_time_names:
+        if time_name in ds.coords:
+            time_coord_name = time_name
+            break
     
-    u10 = ds_instant['u10'].values
-    v10 = ds_instant['v10'].values
+    if time_coord_name is None:
+        raise ValueError(f"Could not find time coordinate. Available coordinates: {list(ds.coords.keys())}")
+    
+    print(f"Using time coordinate: {time_coord_name}")
+    
+    # Get the actual time array from ERA5 data
+    era5_time_array = ds[time_coord_name].values
+    
+    # Convert ERA5 time to datetime objects
+    if hasattr(era5_time_array, 'to_pydatetime'):
+        era5_datetime_array = era5_time_array.to_pydatetime()
+    else:
+        # Handle numpy datetime64
+        era5_datetime_array = [datetime.datetime.fromtimestamp(ts.astype('datetime64[ns]').astype(int) / 1e9) for ts in era5_time_array]
+    
[…]
