### 2025-10-14 f1615e1
Update api_reference.md
