### 2025-10-14 296d399
Update configuration.md
