### 2025-10-14 7c07bb9
Update configuration.md
