### 2025-10-14 7c1df4b
Update configuration.md
