### 2025-10-14 92e05a5
Update configuration.md
