### 2025-10-14 67c8893
Update configuration.md
