### 2025-10-14 95ed0f9
Update configuration.md
