### 2025-10-14 a360102
Update configuration.md
