### 2025-10-14 835028d
Update configuration.md
