### 2025-10-14 ed3e2eb
Update configuration.md
