### 2025-10-14 9edb020
Update configuration.md
