+++ docs/configuration.md
@@ -156,24 +156,8 @@ save_shadow = False   # Shadow maps
 
 **Output Descriptions:**
 
-| Parameter | Output | Description | 
-|-----------|--------|-------------|-----------|
-| `save_tmrt` | Tmrt | Mean Radiant Temperature (°C) 
-| `save_svf` | SVF | Sky View Factor (0-1) 
-| `save_kup` | Kup | Reflected shortwave radiation (W/m²) 
-| `save_kdown` | Kdown | Incoming shortwave radiation (W/m²) 
-| `save_lup` | Lup | Emitted longwave radiation (W/m²) 
-| `save_ldown` | Ldown | Incoming longwave radiation (W/m²) 
-| `save_shadow` | Shadow | Binary shadow maps (0/1) 
-
-!!! tip "Disk Space"
-    Each output is saved as a multi-band GeoTIFF with one band per hour. For a 24-hour simulation of a 1000×1000 pixel tile, expect approximately:
+ Except for sky view factor, each output is saved as a multi-band GeoTIFF with one time step per band.
     
-    - UTCI/Tmrt: ~100 MB per tile
-    - Radiation outputs: ~100 MB each per tile
-    - SVF: ~4 MB per tile (single band)
-    - Shadow: ~25 MB per tile
-
 ## Complete Configuration Example
 
 Here's a complete example showing all configuration options:
