### 2025-10-14 9507f96
Update configuration.md
