+++ solweig_gpu/preprocessor.py
@@ -22,6 +22,7 @@ import netCDF4 as nc
 import xarray as xr
 import shutil
 from netCDF4 import Dataset, date2num
+import datetime as dt
 from datetime import timedelta
 from osgeo import gdal, ogr, osr
 from shapely.geometry import box, Polygon
@@ -185,178 +186,189 @@ def create_tiles(infile, tilesize, overlap, tile_type):
 
 def process_era5_data(start_time, end_time, folder_path, output_file="Outfile.nc"):
     """
-    Process ERA5 NetCDF files to create meteorological forcing data for SOLWEIG.
+    Create a SOLWEIG forcing NetCDF from ERA5 files by slicing the *existing* time range
+    in the inputs [start_time, end_time], inclusive, in UTC.
 
-    Parameters:
-        start_time (str): Start datetime string in format "%Y-%m-%d %H:%M:%S".
-        end_time (str): End datetime string in format "%Y-%m-%d %H:%M:%S".
-        folder_path (str): Path containing ERA5 NetCDF files.
-        output_file (str): Output NetCDF file name.
+    Inputs:
+      - start_time, end_time: "YYYY-MM-DD HH:MM:SS" in UTC (naive strings treated as UTC)
+      - folder_path: contains:
+          data_stream-oper_stepType-instant.nc  (t2m, d2m, sp, u10, v10, ...)
+          data_stream-oper_stepType-accum.nc    (ssrd, strd as accumulated J m^-2)
+    Output:
+      - NetCDF with variables: T2 (°C), PSFC (kPa), RH2 (%), WIND (m/s), SWDOWN (W m^-2)
+        (Add GLW if desired)
     """
-    start_time = datetime.datetime.strptime(start_time, "%Y-%m-%d %H:%M:%S")
-    end_time = datetime.datetime.strptime(end_time, "%Y-%m-%d %H:%M:%S")
-    
-    def saturation_vapor_pressure(T):
+    # --- parse times (treat naive as UTC) ---
+    start_dt = dt.datetime.strptime(start_time, "%Y-%m-%d %H:%M:%S")
+    end_dt   = dt.datetime.strptime(end_time,   "%Y-%m-%d %H:%M:%S")
+
+    instant_file = os.path.join(folder_path, "data_stream-oper_stepType-instant.nc")
+    accum_file   = os.path.join(folder_path, "data_stream-oper_stepType-accum.nc")
+
+    ds_instant = xr.open_dataset(instant_file)
+    ds_accum   = xr.open_dataset(accum_file)
+
+    # --- helper: saturation vapor pressure (Tetens) with T in °C, output hPa ---
+    def esat_c(Tc):
+        return 6.112 * np.exp((17.67 * Tc) / (Tc + 243.5))
+
+    # --- Select the requested time window present in files ---
+    # We'll slice instant variables exactly on [start, end].
+    dsI = ds_instant.sel(time=slice(start_dt, end_dt))
+
+    if dsI.time.size == 0:
+        raise ValueError("No time steps found in 'instant' file within the requested window.")
+
+    # For accumulated variables, to compute hourly rates via differencing,
+    # we need one timestep before the start if available.
+    pre_start = ds_accum.time.sel(time=slice(None, start_dt)).max().values if "time" in ds_accum.coords else None
+    if pre_start is not None and np.isfinite(pre_start):
+        acc_slice_start = dt.datetime.utcfromtimestamp(np.datetime64(pre_start).astype('datetime64[s]').astype(int))
+        # ensure we include one step before start_dt to compute the first diff
+        dsA = ds_accum.sel(time=slice(acc_slice_start, end_dt))
+    else:
+        # no earlier step; we'll slice normally and handle the first step carefully
[…]
