+++ solweig_gpu/preprocessor.py
@@ -188,63 +188,49 @@ def _normalize_time_coord(ds: xr.Dataset) -> xr.Dataset:
     """
     Return a copy of ds with a proper 'time' coordinate.
     Handles common ERA5 layouts:
-      1) 'valid_time' (already-realized timestamps) -> rename to 'time'
-      2) 'time' + 'step' (forecast-style files) -> create time = time + step
-      3) plain 'time' -> leave as-is
+      - 'valid_time' -> rename to 'time'
+      - 'time' + 'step' -> create 'time' = time + step
+      - plain 'time' -> leave as-is
     """
     ds = ds.copy()
 
-    # If 'valid_time' dimension/coord exists, prefer that as time
     if "valid_time" in ds.dims or "valid_time" in ds.coords:
-        ds = ds.rename({"valid_time": "time"})
-        return ds
+        return ds.rename({"valid_time": "time"})
 
-    # If it has CF forecast layout: base 'time' plus 'step' offsets
-    has_time = ("time" in ds.coords) or ("time" in ds.dims)
-    has_step = ("step" in ds.coords) or ("step" in ds.dims)
+    has_time = ("time" in ds.dims) or ("time" in ds.coords)
+    has_step = ("step" in ds.dims) or ("step" in ds.coords)
+
     if has_time and has_step:
-        # Ensure both decode to datetime64 / timedelta64
-        base_time = xr.decode_cf(ds[["time"]])["time"] if "time" in ds.coords else ds["time"]
+        base = ds["time"]
         step = ds["step"]
-        # xarray usually decodes step to timedelta64[h], but if not, try pandas
+
+        # ensure timedelta
         if not np.issubdtype(step.dtype, np.timedelta64):
-            # attempt to parse as hours if it's numeric, else fall back to pandas to_timedelta
             try:
-                step_td = xr.DataArray(
-                    pd.to_timedelta(step.values),
-                    dims=step.dims,
-                    coords=step.coords
-                )
+                step_td = xr.DataArray(pd.to_timedelta(step.values),
+                                       dims=step.dims, coords=step.coords)
             except Exception:
-                # last resort: assume hours
-                step_td = xr.DataArray(
-                    pd.to_timedelta(step.values, unit="h"),
-                    dims=step.dims,
-                    coords=step.coords
-                )
+                # assume hours if not CF-decoded
+                step_td = xr.DataArray(pd.to_timedelta(step.values, unit="h"),
+                                       dims=step.dims, coords=step.coords)
         else:
             step_td = step
 
-        # Broadcast-add base time and step to get valid times
-        vt = (base_time + step_td).rename("time")
-        # If result is 2D (e.g., dims ('time','step')), flatten into a single time axis:
+        vt = (base + step_td).rename("time")
         if vt.ndim == 2:
-            vt = vt.stack(time_flat=("time", "step")).rename("time")
+            # collapse (time, step) → single time axis
+            vt_flat = vt.stack(time_flat=("time", "step")).rename("time")
             ds = ds.stack(time_flat=("time", "step")).rename_dims({"time_flat": "time"}).drop_vars(["time", "step"])
-            ds = ds.assign_coords(time=vt.values)
+            ds = ds.assign_coords(time=vt_flat.values)
         else:
[…]
