+++ solweig_gpu/preprocessor.py
@@ -184,33 +184,33 @@ def create_tiles(infile, tilesize, overlap, tile_type):
 # meteogrological forcing for SOLWEIG. Note that RAIN is set to 0.
 # =============================================================================
 
+import os
+import datetime as dt
+import numpy as np
+import pandas as pd
+import xarray as xr
+from netCDF4 import Dataset, date2num
+
 def _normalize_time_coord(ds: xr.Dataset) -> xr.Dataset:
-    """
-    Return a copy of ds with a proper 'time' coordinate.
-    Handles common ERA5 layouts:
-      - 'valid_time' -> rename to 'time'
-      - 'time' + 'step' -> create 'time' = time + step
-      - plain 'time' -> leave as-is
-    """
+    """Return a copy with a proper 'time' coord from 'valid_time' or 'time'+'step'."""
     ds = ds.copy()
 
+    # Case 1: valid_time -> time
     if "valid_time" in ds.dims or "valid_time" in ds.coords:
         return ds.rename({"valid_time": "time"})
 
+    # Case 2: time + step
     has_time = ("time" in ds.dims) or ("time" in ds.coords)
     has_step = ("step" in ds.dims) or ("step" in ds.coords)
-
     if has_time and has_step:
         base = ds["time"]
         step = ds["step"]
-
-        # ensure timedelta
+        # ensure timedelta dtype
         if not np.issubdtype(step.dtype, np.timedelta64):
             try:
                 step_td = xr.DataArray(pd.to_timedelta(step.values),
                                        dims=step.dims, coords=step.coords)
             except Exception:
-                # assume hours if not CF-decoded
                 step_td = xr.DataArray(pd.to_timedelta(step.values, unit="h"),
                                        dims=step.dims, coords=step.coords)
         else:
@@ -218,7 +218,6 @@ def _normalize_time_coord(ds: xr.Dataset) -> xr.Dataset:
 
         vt = (base + step_td).rename("time")
         if vt.ndim == 2:
-            # collapse (time, step) → single time axis
             vt_flat = vt.stack(time_flat=("time", "step")).rename("time")
             ds = ds.stack(time_flat=("time", "step")).rename_dims({"time_flat": "time"}).drop_vars(["time", "step"])
             ds = ds.assign_coords(time=vt_flat.values)
@@ -226,11 +225,10 @@ def _normalize_time_coord(ds: xr.Dataset) -> xr.Dataset:
             if "step" in vt.dims and "time" not in vt.dims:
                 ds = ds.swap_dims({"step": "time"}).drop_vars(["step"])
             ds = ds.assign_coords(time=vt.values)
-
         return ds
 
+    # Case 3: plain time
     if has_time:
-        # ensure decoding if possible
         try:
             ds = xr.decode_cf(ds)
         except Exception:
@@ -244,127 +242,139 @@ def _normalize_time_coord(ds: xr.Dataset) -> xr.Dataset:
 
 def process_era5_data(start_time, end_time, folder_path, output_file="Outfile.nc"):
     """
-    Process hourly ERA5 NetCDF files to create meteorological forcing for SOLWEIG.
-    Assumes:
-      - Instantaneous fields (t2m, d2m, sp, u10, v10) are hourly
-      - Accumulated fields (ssrd, strd) are 1-hour totals (J m^-2 per hour)
-    Conversion: ssrd/strd → W m^-2 by dividing by 3600 (no differencing).
-
-    Inputs:
-      start_time, end_time: "YYYY-MM-DD HH:MM:SS" (UTC)
-      folder_path: folder with:
-        - data_stream-oper_stepType-instant.nc
[…]
