### 2025-10-14 888d1da
Update configuration.md
