### 2025-10-14 772e27a
Update examples.md
