+++ docs/input_data.md
@@ -8,8 +8,8 @@ All input rasters must be in **GeoTIFF format** and share the same properties to
 
 ### Requirements
 
-!!! "Critical Requirements"
-    All input rasters must have:
+!!! Critical Requirements:
+    All input rasters must have the following 
     
     - Identical Coordinate Reference System (CRS)
     - Same dimensions (width and height in pixels)
@@ -81,42 +81,11 @@ SOLWEIG-GPU supports three different types of meteorological data sources. You m
 
 ### Option 1: Custom Text File (UMEP Format)
 
-A text file with a specific format, typically generated by the UMEP (Urban Multi-scale Environmental Predictor) tool.
-
-**Format:** Space-delimited text file with header
-
-**Required Columns:**
-
-| Column | Description | Units |
-|--------|-------------|-------|
-| Year | Year | - |
-| DOY | Day of year | 1-366 |
-| Hour | Hour of day | 0-23 |
-| Min | Minute | 0-59 |
-| Ta | Air temperature | °C |
-| RH | Relative humidity | % |
-| G | Global radiation | W/m² |
-| D | Diffuse radiation | W/m² |
-| I | Direct radiation | W/m² |
-| Ws | Wind speed | m/s |
-| Wd | Wind direction | degrees |
-| P | Atmospheric pressure | hPa |
-
-**Example:**
-
-```
-Year DOY Hour Min Ta RH G D I Ws Wd P
-2020 225 0 0 22.5 65 0 0 0 2.1 180 1013
-2020 225 1 0 21.8 68 0 0 0 1.9 175 1013
-2020 225 2 0 21.2 70 0 0 0 1.7 170 1013
-```
-
-!!! note "Time Zone"
-    The time in the custom meteorological file should be in local time for the study area.
+A text file with a specific format generated by the UMEP (Urban Multi-scale Environmental Predictor) tool: <https://umep-docs.readthedocs.io/en/latest/pre-processor/Meteorological%20Data%20MetPreprocessor.html>.
 
 ### Option 2: ERA5 Reanalysis Data
 
-NetCDF files from the Copernicus Climate Data Store containing ERA5 reanalysis data.
+NetCDF files from the Copernicus Climate Data Store containing ERA5 reanalysis data: <https://cds.climate.copernicus.eu/datasets/reanalysis-era5-single-levels?tab=overview>.
 
 **Required Files:** Both instantaneous and accumulated data files
 
@@ -132,13 +101,7 @@ NetCDF files from the Copernicus Climate Data Store containing ERA5 reanalysis d
 | ssrd | Surface solar radiation downwards (accumulated) | J/m² |
 | strd | Surface thermal radiation downwards (accumulated) | J/m² |
 
-**File Naming:** ERA5 files should be named according to the standard ERA5 naming convention.
-
-!!! info "Downloading ERA5 Data"
-    ERA5 data can be downloaded from the [Copernicus Climate Data Store](https://cds.climate.copernicus.eu/). You will need to register for a free account.
-
-!!! warning "Time Specification"
-    When using ERA5 data, you must specify `start_time` and `end_time` in **UTC**. The package will automatically convert to local time based on the location of your study area.
+**File Naming:** ERA5 files should be named according to the standard ERA5 naming convention (do not rename the downloaded ERA-5 instantaneous and accumulated data).
 
 ### Option 3: WRF Output Files
 
@@ -165,24 +128,6 @@ Where `X` is the domain number (1-9).
 
 Choose an appropriate projected coordinate system for your study area. UTM zones are commonly used for urban-scale studies.
 
-!!! tip "CRS Selection"
-    - Use a projected CRS (not geographic lat/lon)
-    - UTM zones work well for most applications
-    - Ensure all rasters use the same CRS
-
[…]
