### 2025-10-15 dfe0cf7
Update configuration.md
