### 2025-10-15 973c599
Update configuration.md
