+++ docs/configuration.md
@@ -131,14 +131,14 @@ end_time = '2020-08-13 23:00:00'    # UTC
 use_own_met = False
 data_source_type = 'ERA5'
 data_folder = '/path/to/era5_data'
-start_time = '2020-08-13 06:00:00'  # UTC
-end_time = '2020-08-14 05:00:00'    # UTC
+start_time = '2020-08-13 00:00:00'  # UTC
+end_time = '2020-08-14 23:00:00'    # UTC
 ```
 
 !!! Important Notes:
    1.  When using ERA5 or WRF data, `start_time` and `end_time` must be in **UTC**. The package will automatically convert to local time based on the geographic location of your study area.
    2.  When using wrfout, the `start_time` and `end_time` should be the first and last time stamps in the wrfout dataset. The package won't compare the selecetd `start_time` and `end_time` to the wrfout file timestamps and automatically fetch the corresponding data.
-   3.  When using ERA-5 dataset, the package can find the corresponding data for `start_time` and `end_time`. For example, if ERA-5 data is downloaded from 2020-08-13 00 UTC to 2020-08-14 23 UTC and the model is to be run from 2020-08-13 06 UTC to 2020-08-14 05 UTC, the package can select data from 2020-08-13 06 UTC to 2020-08-14 05 UTC (selected_date_str = '2020-08-13', start_time = '2020-08-13 06:00:00', and end_time = '2020-08-14 05:00:00')
+   3.  When using ERA-5 dataset, the package can find the corresponding data for `start_time` and `end_time`. For example, if ERA-5 data is downloaded from 2020-08-13 00 UTC to 2020-08-14 23 UTC and the model is to be run from 2020-08-13 06 UTC to 2020-08-14 05 UTC, the package can select data from 2020-08-13 06 UTC to 2020-08-14 05 UTC by itself (selected_date_str = '2020-08-13', start_time = '2020-08-13 00:00:00', and end_time = '2020-08-14 23:00:00')
     
 
 ## Output Configuration
