### 2025-10-15 01f7765
Update outputs.md
