### 2025-10-15 d56bcf6
Update documentation links for input and configuration
