### 2025-10-15 8584402
Add link to Jupyter notebook in examples.md
Added a reference to the Jupyter notebook for ERA5 examples.
