### 2025-10-15 c315c08
Add Jupyter notebook links to examples
Added links to Jupyter notebooks for examples using WRF and own meteorological data.
