### 2025-10-15 34b4868
Change simulation time range for ERA5
Updated start and end times for the simulation.
