+++ tests/test_integration_end_to_end.py
@@ -1,0 +1,73 @@
+import os
+import shutil
+import tempfile
+import unittest
+import numpy as np
+from osgeo import gdal, osr
+
+
+def _create_raster(path: str, width=32, height=32, pixel_size=0.0005, epsg=4326):
+    driver = gdal.GetDriverByName('GTiff')
+    ds = driver.Create(path, width, height, 1, gdal.GDT_Float32)
+    geotransform = (77.0, pixel_size, 0.0, 29.25, 0.0, -pixel_size)
+    ds.SetGeoTransform(geotransform)
+    srs = osr.SpatialReference(); srs.ImportFromEPSG(epsg)
+    ds.SetProjection(srs.ExportToWkt())
+    band = ds.GetRasterBand(1)
+    band.WriteArray(np.random.rand(height, width).astype(np.float32))
+    ds.FlushCache(); ds = None
+    return path
+
+
+class TestEndToEnd(unittest.TestCase):
+    def setUp(self):
+        self.temp_dir = tempfile.mkdtemp()
+        self.input_dir = self.temp_dir
+        # Create required rasters
+        _create_raster(os.path.join(self.input_dir, 'Building_DSM.tif'))
+        _create_raster(os.path.join(self.input_dir, 'DEM.tif'))
+        _create_raster(os.path.join(self.input_dir, 'Trees.tif'))
+
+        # Create a trivial metfile (own met path)
+        self.metfile = os.path.join(self.input_dir, 'met.txt')
+        with open(self.metfile, 'w') as f:
+            f.write('iy id it imin Q* QH QE Qs Qf Wind RH Td press rain Kdn snow ldown fcld wuh xsmd lai_hr Kdiff Kdir Wd\n')
+            for h in range(24):
+                f.write(f"2020 200 {h} 0 -999 -999 -999 -999 -999 1.0 50 20 100 0 0 -999 -999 -999 -999 -999 -999 -999 -999 -999\n")
+
+    def tearDown(self):
+        shutil.rmtree(self.temp_dir, ignore_errors=True)
+
+    def test_pipeline_with_own_met(self):
+        # Use own met path to avoid depending on external datasets
+        from solweig_gpu import thermal_comfort
+        thermal_comfort(
+            base_path=self.input_dir,
+            selected_date_str='2020-07-18',
+            building_dsm_filename='Building_DSM.tif',
+            dem_filename='DEM.tif',
+            trees_filename='Trees.tif',
+            landcover_filename=None,
+            tile_size=32,
+            overlap=0,
+            use_own_met=True,
+            start_time='2020-07-18 00:00:00',
+            end_time='2020-07-18 23:00:00',
+            data_source_type='ERA5',
+            data_folder=self.input_dir,
+            own_met_file=self.metfile,
+            save_svf=False,
+        )
+
+        # Verify outputs exist
+        outputs_dir = os.path.join(self.input_dir, 'Outputs')
+        self.assertTrue(os.path.isdir(outputs_dir))
+        # There should be at least one tile folder with results
+        tile_dirs = [d for d in os.listdir(outputs_dir) if os.path.isdir(os.path.join(outputs_dir, d))]
+        self.assertTrue(len(tile_dirs) >= 1)
+
+
+if __name__ == '__main__':
+    unittest.main()
+
+
