### 2025-10-15 a005518
adding tests
