+++ tests/test_era5_processing.py
@@ -1,0 +1,177 @@
+"""
+Tests for ERA5 processing: multi-file loading and time coordinate handling.
+"""
+
+import os
+import shutil
+import tempfile
+import unittest
+import numpy as np
+import xarray as xr
+from netCDF4 import Dataset
+from osgeo import gdal, osr
+
+
+def _create_tiny_dem_tile(folder_path: str, filename: str,
+                          origin_x: float, origin_y: float,
+                          pixel_size: float = 0.01,
+                          width: int = 10, height: int = 10) -> str:
+    if not os.path.exists(folder_path):
+        os.makedirs(folder_path)
+    driver = gdal.GetDriverByName('GTiff')
+    filepath = os.path.join(folder_path, filename)
+    ds = driver.Create(filepath, width, height, 1, gdal.GDT_Float32)
+    geotransform = (origin_x, pixel_size, 0.0, origin_y, 0.0, -pixel_size)
+    ds.SetGeoTransform(geotransform)
+    srs = osr.SpatialReference()
+    srs.ImportFromEPSG(4326)
+    ds.SetProjection(srs.ExportToWkt())
+    band = ds.GetRasterBand(1)
+    data = np.ones((height, width), dtype=np.float32)
+    band.WriteArray(data)
+    ds.FlushCache()
+    ds = None
+    return filepath
+
+
+class TestERA5Processing(unittest.TestCase):
+    def setUp(self):
+        self.temp_dir = tempfile.mkdtemp()
+        self.era5_dir = os.path.join(self.temp_dir, "era5")
+        os.makedirs(self.era5_dir, exist_ok=True)
+
+        # Create synthetic ERA5-like datasets split across two files
+        # Grid 2x3 centered over a plausible location
+        lat = np.array([29.0, 29.25], dtype=np.float32)
+        lon = np.array([77.0, 77.25, 77.5], dtype=np.float32)
+
+        # 24 hourly timesteps on a specific UTC date
+        times_all = np.array(
+            np.arange('2000-08-13T00:00', '2000-08-14T00:00', dtype='datetime64[h]')
+        )
+        # Split into two halves to test open_mfdataset
+        times_1 = times_all[:12]
+        times_2 = times_all[12:]
+
+        def make_vars(time_len):
+            shape = (time_len, lat.size, lon.size)
+            return {
+                't2m': (('valid_time', 'latitude', 'longitude'), 300.0 + np.zeros(shape, dtype=np.float32)),
+                'd2m': (('valid_time', 'latitude', 'longitude'), 295.0 + np.zeros(shape, dtype=np.float32)),
+                'sp': (('valid_time', 'latitude', 'longitude'), 100000.0 + np.zeros(shape, dtype=np.float32)),
+                'u10': (('valid_time', 'latitude', 'longitude'), 2.0 + np.zeros(shape, dtype=np.float32)),
+                'v10': (('valid_time', 'latitude', 'longitude'), 1.0 + np.zeros(shape, dtype=np.float32)),
+                'ssrd': (('valid_time', 'latitude', 'longitude'), 3600.0 + np.zeros(shape, dtype=np.float32)),
+                'strd': (('valid_time', 'latitude', 'longitude'), 3600.0 + np.zeros(shape, dtype=np.float32)),
+            }
+
+        ds1 = xr.Dataset(
+            data_vars=make_vars(len(times_1)),
+            coords={
+                'valid_time': times_1,
+                'latitude': lat,
+                'longitude': lon,
+            },
+            attrs={'Conventions': 'CF-1.7'}
+        )
+        ds2 = xr.Dataset(
+            data_vars=make_vars(len(times_2)),
+            coords={
+                'valid_time': times_2,
+                'latitude': lat,
[…]
