+++ .github/workflows/tests.yml
@@ -29,6 +29,13 @@ jobs:
         if: runner.os == 'Linux'
         run: sudo apt-get update && sudo apt-get install -y gdal-bin libgdal-dev
 
+      - name: Install system deps (GDAL - macOS)
+        if: runner.os == 'macOS'
+        run: |
+          brew update
+          brew install gdal
+          echo "GDAL installed via Homebrew"
+
       - name: Install package
         run: |
           python -m pip install --upgrade pip
