+++ .github/workflows/tests.yml
@@ -27,7 +27,10 @@ jobs:
 
       - name: Install system deps (GDAL)
         if: runner.os == 'Linux'
-        run: sudo apt-get update && sudo apt-get install -y gdal-bin libgdal-dev
+        run: |
+          sudo add-apt-repository -y ppa:ubuntugis/ppa
+          sudo apt-get update
+          sudo apt-get install -y gdal-bin libgdal-dev
 
       - name: Install system deps (GDAL - macOS)
         if: runner.os == 'macOS'
@@ -39,8 +42,8 @@ jobs:
       - name: Install package
         run: |
           python -m pip install --upgrade pip
-          pip install -e .
           pip install -r requirements.txt || true
+          pip install -e . --no-build-isolation
           pip install pytest coverage pytest-cov
 
       - name: Run tests with coverage
