+++ .github/workflows/tests.yml
@@ -20,33 +20,24 @@ jobs:
       - name: Checkout
         uses: actions/checkout@v4
 
-      - name: Set up Python
-        uses: actions/setup-python@v5
+      - name: Setup Mambaforge
+        uses: conda-incubator/setup-miniconda@v3
         with:
+          miniforge-variant: Mambaforge
+          miniforge-version: latest
+          activate-environment: test
           python-version: ${{ matrix.python-version }}
+          use-mamba: true
 
-      - name: Install system deps (GDAL)
-        if: runner.os == 'Linux'
+      - name: Install dependencies
+        shell: bash -el {0}
         run: |
-          sudo add-apt-repository -y ppa:ubuntugis/ppa
-          sudo apt-get update
-          sudo apt-get install -y gdal-bin libgdal-dev
-
-      - name: Install system deps (GDAL - macOS)
-        if: runner.os == 'macOS'
-        run: |
-          brew update
-          brew install gdal
-          echo "GDAL installed via Homebrew"
-
-      - name: Install package
-        run: |
-          python -m pip install --upgrade pip
-          pip install -r requirements.txt || true
-          pip install -e . --no-build-isolation
+          mamba install -c conda-forge gdal numpy scipy pandas netcdf4 pytz shapely timezonefinder xarray tqdm matplotlib-base pyqt
           pip install pytest coverage pytest-cov
+          pip install -e . --no-deps
 
       - name: Run tests with coverage
+        shell: bash -el {0}
         run: |
           pytest --cov=solweig_gpu --cov-report=xml --cov-report=term-missing
 
