### 2025-10-15 35808b5
Update test_gpu_smoke.py
