+++ .github/workflows/tests.yml
@@ -13,8 +13,9 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        os: [ubuntu-latest, macos-latest, windows-latest]
+        os: [ubuntu-latest, macos-latest]
         python-version: ["3.10", "3.11", "3.12"]
+        # Windows excluded due to PyTorch DLL compatibility issues in CI
 
     steps:
       - name: Checkout
