+++ tests/test_calculate_utci.py
@@ -13,41 +13,106 @@ class TestUTCICalculation(unittest.TestCase):
 
     def test_utci_comfortable_conditions(self):
         """Test UTCI calculation for comfortable conditions."""
-        # This is a placeholder test
-        # Actual implementation would test the utci_calculator function
-        # with known reference values
+        from solweig_gpu.calculate_utci import utci_calculator
         
-        # Example comfortable conditions:
-        # Ta = 25°C, Tmrt = 25°C, wind = 1 m/s, RH = 50%
+        # Comfortable conditions: Ta = 25°C, Tmrt = 25°C, wind = 1 m/s, RH = 50%
         # Expected UTCI should be close to 25°C
-        pass
+        ta = 25.0  # Air temperature
+        tmrt = 25.0  # Mean radiant temperature
+        wind = 1.0  # Wind speed
+        rh = 50.0  # Relative humidity
+        
+        utci = utci_calculator(ta, tmrt, wind, rh)
+        
+        # UTCI should be close to air temp in comfortable conditions
+        self.assertIsInstance(utci, (int, float, np.ndarray))
+        self.assertTrue(23.0 < utci < 27.0, f"UTCI {utci} out of expected range")
 
     def test_utci_hot_conditions(self):
         """Test UTCI calculation for hot conditions."""
-        # Test with high temperature and high radiation
-        # Expected UTCI should be significantly higher than air temperature
-        pass
+        from solweig_gpu.calculate_utci import utci_calculator
+        
+        # Hot conditions: high temperature and high radiation
+        ta = 35.0
+        tmrt = 50.0  # High radiant temperature
+        wind = 0.5
+        rh = 40.0
+        
+        utci = utci_calculator(ta, tmrt, wind, rh)
+        
+        # UTCI should be higher than air temperature due to radiation
+        self.assertIsInstance(utci, (int, float, np.ndarray))
+        self.assertTrue(utci > ta, f"UTCI {utci} should be > air temp {ta}")
 
     def test_utci_cold_conditions(self):
         """Test UTCI calculation for cold conditions."""
-        # Test with low temperature and high wind
-        # Expected UTCI should be lower than air temperature (wind chill)
-        pass
+        from solweig_gpu.calculate_utci import utci_calculator
+        
+        # Cold conditions: low temperature and high wind (wind chill)
+        ta = 5.0
+        tmrt = 5.0
+        wind = 5.0  # High wind
+        rh = 70.0
+        
+        utci = utci_calculator(ta, tmrt, wind, rh)
+        
+        # UTCI should be lower than air temperature due to wind chill
+        self.assertIsInstance(utci, (int, float, np.ndarray))
+        self.assertTrue(utci < ta, f"UTCI {utci} should be < air temp {ta} due to wind")
 
     def test_utci_extreme_heat(self):
         """Test UTCI calculation for extreme heat conditions."""
-        # Test boundary conditions for extreme heat
-        pass
+        from solweig_gpu.calculate_utci import utci_calculator
+        
+        # Extreme heat
+        ta = 40.0
+        tmrt = 60.0
+        wind = 0.3
+        rh = 30.0
+        
+        utci = utci_calculator(ta, tmrt, wind, rh)
+        
+        # Should return a valid number (not NaN or inf)
+        self.assertIsInstance(utci, (int, float, np.ndarray))
+        self.assertFalse(np.isnan(utci))
+        self.assertFalse(np.isinf(utci))
 
     def test_utci_extreme_cold(self):
         """Test UTCI calculation for extreme cold conditions."""
-        # Test boundary conditions for extreme cold
-        pass
[…]

[…]
