+++ tests/test_calculate_utci.py
@@ -6,173 +6,216 @@ This module tests the UTCI (Universal Thermal Climate Index) calculation.
 
 import unittest
 import numpy as np
+import pytest
+
+try:
+    import torch
+    TORCH_AVAILABLE = True
+except ImportError:
+    TORCH_AVAILABLE = False
 
 
 class TestUTCICalculation(unittest.TestCase):
     """Test UTCI calculation functionality."""
 
+    @pytest.mark.skipif(not TORCH_AVAILABLE, reason='PyTorch required for UTCI')
     def test_utci_comfortable_conditions(self):
         """Test UTCI calculation for comfortable conditions."""
         from solweig_gpu.calculate_utci import utci_calculator
+        import torch
         
         # Comfortable conditions: Ta = 25°C, Tmrt = 25°C, wind = 1 m/s, RH = 50%
         # Expected UTCI should be close to 25°C
-        ta = 25.0  # Air temperature
-        tmrt = 25.0  # Mean radiant temperature
-        wind = 1.0  # Wind speed
-        rh = 50.0  # Relative humidity
+        # Signature: utci_calculator(Ta, RH, Tmrt, va10m)
+        ta = torch.tensor([25.0], dtype=torch.float32)
+        rh = torch.tensor([50.0], dtype=torch.float32)
+        tmrt = torch.tensor([25.0], dtype=torch.float32)
+        wind = torch.tensor([1.0], dtype=torch.float32)
         
-        utci = utci_calculator(ta, tmrt, wind, rh)
+        utci = utci_calculator(ta, rh, tmrt, wind)
         
         # UTCI should be close to air temp in comfortable conditions
-        self.assertIsInstance(utci, (int, float, np.ndarray))
-        self.assertTrue(23.0 < utci < 27.0, f"UTCI {utci} out of expected range")
+        utci_val = utci.item() if hasattr(utci, 'item') else float(utci[0])
+        self.assertTrue(23.0 < utci_val < 27.0, f"UTCI {utci_val} out of expected range")
 
+    @pytest.mark.skipif(not TORCH_AVAILABLE, reason='PyTorch required for UTCI')
     def test_utci_hot_conditions(self):
         """Test UTCI calculation for hot conditions."""
         from solweig_gpu.calculate_utci import utci_calculator
+        import torch
         
         # Hot conditions: high temperature and high radiation
-        ta = 35.0
-        tmrt = 50.0  # High radiant temperature
-        wind = 0.5
-        rh = 40.0
+        # Signature: utci_calculator(Ta, RH, Tmrt, va10m)
+        ta = torch.tensor([35.0], dtype=torch.float32)
+        rh = torch.tensor([40.0], dtype=torch.float32)
+        tmrt = torch.tensor([50.0], dtype=torch.float32)
+        wind = torch.tensor([0.5], dtype=torch.float32)
         
-        utci = utci_calculator(ta, tmrt, wind, rh)
+        utci = utci_calculator(ta, rh, tmrt, wind)
         
         # UTCI should be higher than air temperature due to radiation
-        self.assertIsInstance(utci, (int, float, np.ndarray))
-        self.assertTrue(utci > ta, f"UTCI {utci} should be > air temp {ta}")
+        utci_val = utci.item() if hasattr(utci, 'item') else float(utci[0])
+        self.assertTrue(utci_val > 35.0, f"UTCI {utci_val} should be > air temp 35.0")
 
+    @pytest.mark.skipif(not TORCH_AVAILABLE, reason='PyTorch required for UTCI')
     def test_utci_cold_conditions(self):
         """Test UTCI calculation for cold conditions."""
         from solweig_gpu.calculate_utci import utci_calculator
+        import torch
         
[…]
