### 2025-10-15 65e8bda
Refactor installation instructions in README
Updated installation instructions to avoid repetition.
