### 2025-10-15 87b727a
Simplify contributing section in README
Removed detailed contributing steps from README.
