### 2025-10-15 505ea0c
Restore badges for tests and code coverage
Re-added test and code coverage badges to the README.
