### 2025-10-16 9724d80
Update README.md with new tests badge
Removed the old tests badge and added a new tests workflow badge.
