### 2025-10-16 4b955d6
Delete GITHUB_PAGES_SETUP.md
