+++ GITHUB_PAGES_SETUP.md
@@ -1,222 +1,0 @@
-# GitHub Pages Setup Guide for SOLWEIG-GPU
-
-This guide explains how to set up and deploy the SOLWEIG-GPU documentation to GitHub Pages.
-
-## Overview
-
-The documentation is built using **MkDocs** with the **Material for MkDocs** theme. It is automatically deployed to GitHub Pages using GitHub Actions whenever changes are pushed to the `main` branch.
-
-## Prerequisites
-
-Before setting up GitHub Pages, ensure you have:
-
-1. A GitHub repository for SOLWEIG-GPU
-2. Admin access to the repository
-3. The documentation files in the `docs/` directory
-4. The `mkdocs.yml` configuration file in the repository root
-
-## Setup Steps
-
-### Step 1: Enable GitHub Pages
-
-1. Go to your repository on GitHub
-2. Click on **Settings** (in the repository menu)
-3. Scroll down to the **Pages** section in the left sidebar
-4. Under **Source**, select **Deploy from a branch**
-5. Select the `gh-pages` branch and `/ (root)` folder
-6. Click **Save**
-
-### Step 2: Configure Repository Permissions
-
-The GitHub Actions workflow needs permission to deploy to GitHub Pages:
-
-1. In **Settings**, go to **Actions** → **General**
-2. Scroll down to **Workflow permissions**
-3. Select **Read and write permissions**
-4. Check **Allow GitHub Actions to create and approve pull requests**
-5. Click **Save**
-
-### Step 3: Push the Documentation Files
-
-Make sure all the documentation files are committed and pushed to your repository:
-
-```bash
-git add docs/ mkdocs.yml .github/workflows/docs.yml
-git commit -m "Add documentation and GitHub Pages deployment"
-git push origin main
-```
-
-### Step 4: Verify Deployment
-
-1. Go to the **Actions** tab in your repository
-2. You should see a workflow run called "Deploy Documentation"
-3. Wait for the workflow to complete (it should take 1-2 minutes)
-4. Once complete, your documentation will be available at:
-   ```
-   https://<username>.github.io/<repository-name>/
-   ```
-
-For the SOLWEIG-GPU repository, this would be:
-```
-https://nvnsudharsan.github.io/SOLWEIG-GPU/
-```
-
-## Local Development
-
-To preview the documentation locally before pushing:
-
-### Install MkDocs
-
-```bash
-pip install mkdocs-material mkdocstrings[python] pymdown-extensions
-```
-
-### Preview Locally
-
-```bash
-cd /path/to/SOLWEIG-GPU
-mkdocs serve
-```
-
-This will start a local server at `http://127.0.0.1:8000/` where you can preview the documentation.
-
-### Build Locally
-
-To build the documentation without serving it:
-
-```bash
-mkdocs build
-```
-
-This creates a `site/` directory with the static HTML files.
-
-## Customization
-
-### Updating the Theme
-
-The documentation uses the Material for MkDocs theme. You can customize it by editing `mkdocs.yml`:
-
-```yaml
-theme:
-  name: material
-  palette:
-    primary: teal
-    accent: amber
-  features:
-    - navigation.tabs
-    - navigation.sections
-    - search.highlight
-```
-
-See the [Material for MkDocs documentation](https://squidfunk.github.io/mkdocs-material/) for more customization options.
-
-### Adding Pages
-
-To add a new page to the documentation:
-
-1. Create a new Markdown file in the `docs/` directory
-2. Add it to the navigation in `mkdocs.yml`:
-
-```yaml
-nav:
-  - Home: index.md
-  - Your New Page: new_page.md
-```
-
-### Adding Images
-
[…]
