+++ DOCUMENTATION_README.md
@@ -1,276 +1,0 @@
-# SOLWEIG-GPU Documentation, Tests, and Examples
-
-This package includes comprehensive documentation, a test suite, and usage examples for SOLWEIG-GPU.
-
-## What's Included
-
-### 📚 Documentation (GitHub Pages Ready)
-
-A complete documentation website built with MkDocs and the Material theme, ready to be deployed to GitHub Pages.
-
-**Documentation Structure:**
-
-- **Home** (`docs/index.md`): Welcome page with overview and quick links
-- **Getting Started**:
-  - Installation guide with prerequisites and step-by-step instructions
-  - Quickstart guide for running your first simulation
-- **User Guide**:
-  - Input Data: Detailed information on preparing geospatial rasters and meteorological data
-  - Configuration: Complete parameter reference and optimization tips
-  - Outputs: Guide to understanding and visualizing results
-- **API Reference** (`docs/api_reference.md`): Detailed documentation of the `thermal_comfort()` function
-- **Examples** (`docs/examples.md`): Usage examples for different meteorological data sources
-- **Developer Guide** (`docs/developer_guide.md`): Architecture overview and contribution guidelines
-- **Testing** (`docs/testing.md`): Guide to running and writing tests
-
-**Key Features:**
-
-- Modern, responsive design with Material for MkDocs theme
-- Dark/light mode toggle
-- Full-text search
-- Code syntax highlighting
-- Admonitions (notes, tips, warnings)
-- Mobile-friendly navigation
-- Automatic deployment via GitHub Actions
-
-### 🧪 Test Suite
-
-A comprehensive test framework using pytest, organized into multiple modules:
-
-**Test Modules:**
-
-- `test_preprocessor.py`: Tests for raster validation, tiling, and meteorological data extraction
-- `test_sun_position.py`: Tests for solar position and day length calculations
-- `test_calculate_utci.py`: Tests for UTCI calculation accuracy
-- `test_walls_aspect.py`: Tests for wall height and aspect calculations
-- `test_integration.py`: End-to-end integration tests
-
-**Test Configuration:**
-
-- `pytest.ini`: Test configuration with markers for slow, integration, and GPU tests
-- `tests/README.md`: Testing guide with instructions
-
-**Current Status:** Test structure is in place with some implemented tests and many placeholders ready for full implementation.
-
-### 💡 Examples
-
-Practical example scripts demonstrating different use cases:
-
-- `run_with_wrf.py`: Using WRF meteorological data
-- `run_with_era5.py`: Using ERA5 reanalysis data
-- `run_with_custom_met_file.py`: Using custom meteorological files
-- `examples/README.md`: Instructions for running examples
-
-## Setting Up GitHub Pages
-
-### Quick Start
-
-1. **Push to GitHub:**
-
-   ```bash
-   git add .
-   git commit -m "Add documentation, tests, and examples"
-   git push origin main
-   ```
-
-2. **Enable GitHub Pages:**
-
-   - Go to repository **Settings** → **Pages**
-   - Source: Deploy from a branch
-   - Branch: `gh-pages` / `/ (root)`
-   - Click **Save**
-
-3. **Configure Workflow Permissions:**
-
-   - Settings → **Actions** → **General**
-   - Workflow permissions: **Read and write permissions**
-   - Check **Allow GitHub Actions to create and approve pull requests**
-   - Click **Save**
-
-4. **Wait for Deployment:**
-
-   - Go to **Actions** tab
-   - Wait for "Deploy Documentation" workflow to complete
-   - Documentation will be available at: `https://<username>.github.io/<repository>/`
-
[…]
