### 2025-10-16 57a5133
Delete DOCUMENTATION_OVERVIEW.md
