### 2025-10-16 e805ed9
add docstrings
