### 2025-10-16 00399b6
add docstrings
