+++ solweig_gpu/sun_position.py
@@ -24,7 +24,44 @@ from scipy.ndimage import rotate
 
 
 def sun_position(time, location):
-
+    """
+    Calculate solar position (zenith and azimuth) using SPA algorithm.
+    
+    Implements the Solar Position Algorithm (SPA) as described in:
+    Reda, I. and Andreas, A. (2004). Solar position algorithm for solar radiation applications.
+    Solar Energy, 76(5), 577-589.
+    
+    Args:
+        time (dict): Time information with keys:
+            - 'year' (int): Year
+            - 'month' (int): Month (1-12)
+            - 'day' (int): Day of month
+            - 'hour' (int): Hour (0-23)
+            - 'min' (int): Minute (0-59)
+            - 'sec' (int): Second (0-59)
+            - 'UTC' (float): UTC offset in hours (e.g., -5 for EST)
+        
+        location (dict): Geographic location with keys:
+            - 'latitude' (float): Latitude in degrees (-90 to 90)
+            - 'longitude' (float): Longitude in degrees (-180 to 180)
+            - 'altitude' (float): Elevation above sea level in meters
+    
+    Returns:
+        dict: Solar position containing:
+            - 'zenith' (float): Solar zenith angle in degrees (0=directly overhead, 90=horizon)
+            - 'azimuth' (float): Solar azimuth in degrees (0=North, 90=East, 180=South, 270=West)
+    
+    Notes:
+        - Accounts for atmospheric refraction
+        - Accuracy: ~0.0003° for years 2000-6000
+        - All angles in degrees unless otherwise specified
+    
+    Example:
+        >>> time = {'year': 2020, 'month': 7, 'day': 18, 'hour': 12, 'min': 0, 'sec': 0, 'UTC': -5}
+        >>> location = {'latitude': 30.27, 'longitude': -97.74, 'altitude': 0}
+        >>> sun = sun_position(time, location)
+        >>> print(f"Zenith: {sun['zenith']:.2f}°, Azimuth: {sun['azimuth']:.2f}°")
+    """
 
     # 1. Calculate the Julian Day, and Century. Julian Ephemeris day, century
     # and millenium are calculated using a mean delta_t of 33.184 seconds.
@@ -80,6 +117,15 @@ def sun_position(time, location):
 
 
 def julian_calculation(t_input):
+    """
+    Calculate Julian day and related time parameters.
+    
+    Args:
+        t_input (dict): Time dictionary with year, month, day, hour, min, sec, UTC
+    
+    Returns:
+        dict: Julian day, century, ephemeris day/century/millennium
+    """
 
     if not isinstance(t_input, dict):
 
@@ -141,6 +187,15 @@ def julian_calculation(t_input):
 
 
 def earth_heliocentric_position_calculation(julian):
+    """
+    Calculate Earth's heliocentric position (longitude, latitude, radius).
+    
+    Args:
+        julian (dict): Julian day parameters
+    
+    Returns:
+        dict: Earth heliocentric position (longitude, latitude, radius in AU)
+    """
 
     # Tabulated values for the longitude calculation
     # L terms  from the original code.
@@ -457,6 +512,7 @@ def earth_heliocentric_position_calculation(julian):
 
 
 def sun_geocentric_position_calculation(earth_heliocentric_position):
+    """Calculate geocentric sun position from Earth heliocentric position. SPA Step 3."""
 
     sun_geocentric_position = dict()
     sun_geocentric_position['longitude'] = earth_heliocentric_position['longitude'] + 180
@@ -470,6 +526,15 @@ def sun_geocentric_position_calculation(earth_heliocentric_position):
 
 
 def nutation_calculation(julian):
+    """
[…]
