+++ solweig_gpu/solweig.py
@@ -28,6 +28,16 @@ from .shadow import create_patches
 gdal.UseExceptions()
 
 def ensure_tensor(x, device=None):
+    """
+    Convert input to PyTorch tensor on specified device.
+    
+    Args:
+        x: Input data (numpy array, list, or tensor)
+        device (torch.device, optional): Target device
+    
+    Returns:
+        torch.Tensor: Input as tensor on device
+    """
     if device is None:
         device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
     if not isinstance(x, torch.Tensor):
@@ -35,6 +45,20 @@ def ensure_tensor(x, device=None):
     return x
 
 def daylen(DOY, XLAT):
+    """
+    Calculate day length and solar declination for given day and latitude.
+    
+    Args:
+        DOY (torch.Tensor): Day of year (1-365)
+        XLAT (torch.Tensor): Latitude in degrees
+    
+    Returns:
+        tuple: (DAYL, DEC, SNDN, SNUP) where:
+            - DAYL: Day length in hours
+            - DEC: Solar declination in degrees
+            - SNDN: Time of solar noon in hours
+            - SNUP: Time of sunrise in hours
+    """
     RAD = torch.pi / 180.0
 
     DEC = -23.45 * torch.cos(2.0 * torch.pi * (DOY + 10.0) / 365.0)
@@ -50,6 +74,37 @@ def daylen(DOY, XLAT):
 
 def sunonsurface_2018a(azimuthA, scale, buildings, shadow, sunwall, first, second, aspect, walls, Tg, Tgwall, Ta,
                        emis_grid, ewall, alb_grid, SBC, albedo_b, Twater, lc_grid, landcover):
+    """
+    Calculate solar radiation on surfaces with different orientations.
+    
+    Determines radiation on walls and ground surfaces accounting for
+    building geometry, shadows, and surface properties.
+    
+    Args:
+        azimuthA (float): Solar azimuth angle (degrees)
+        scale (float): Grid scale (pixels per meter)
+        buildings (torch.Tensor): Building mask array
+        shadow (torch.Tensor): Shadow map
+        sunwall (torch.Tensor): Sunlit wall mask
+        first (torch.Tensor): First surface type
+        second (torch.Tensor): Second surface type
+        aspect (torch.Tensor): Wall aspect angles
+        walls (torch.Tensor): Wall heights
+        Tg (torch.Tensor): Ground temperature
+        Tgwall (torch.Tensor): Wall temperature
+        Ta (float): Air temperature
+        emis_grid (torch.Tensor): Ground emissivity
+        ewall (float): Wall emissivity
+        alb_grid (torch.Tensor): Ground albedo
+        SBC (float): Stefan-Boltzmann constant
+        albedo_b (float): Building albedo
+        Twater (float): Water temperature
+        lc_grid (torch.Tensor): Land cover grid
+        landcover (np.ndarray): Land cover classification
+    
+    Returns:
+        tuple: Radiation components for different surfaces
+    """
 
     device = torch.device('cuda' if torch.cuda.is_available() else 'cpu')
 
@@ -235,6 +290,34 @@ def sunonsurface_2018a(azimuthA, scale, buildings, shadow, sunwall, first, secon
 
 def gvf_2018a(wallsun, walls, buildings, scale, shadow, first, second, dirwalls, Tg, Tgwall, Ta, emis_grid, ewall,
               alb_grid, SBC, albedo_b, rows, cols, Twater, lc_grid, landcover):
+    """
+    Calculate ground view factors for radiation exchange between surfaces.
+    
[…]
