+++ docs/requirements.txt
@@ -1,4 +1,5 @@
-mkdocs>=1.5.0
-mkdocs-material>=9.0.0
-mkdocstrings[python]>=0.24.0
-pymdown-extensions>=10.0
+# Sphinx documentation dependencies
+sphinx>=5.0
+sphinx-rtd-theme>=1.0
+myst-parser>=0.18
+sphinx-autodoc-typehints>=1.19
