### 2025-10-16 c6c61c9
fix the gdal problem in readthedocs
