+++ .readthedocs.yml
@@ -30,8 +30,4 @@ build:
 python:
   install:
     - requirements: docs/requirements.txt
-    - method: pip
-      path: .
-      extra_requirements:
-        - docs
 
