+++ docs/requirements.txt
@@ -3,3 +3,4 @@ sphinx>=5.0
 sphinx-rtd-theme>=1.0
 myst-parser>=0.18
 sphinx-autodoc-typehints>=1.19
+linkify-it-py>=2.0.0
