+++ docs/index.md
@@ -1,102 +1,0 @@
-# Welcome to SOLWEIG-GPU
-
-<p align="center">
-  <img src="https://raw.githubusercontent.com/nvnsudharsan/solweig-gpu/main/Logo_solweig.jpg" alt="SOLWEIG Logo" width="400"/>
-</p>
-<p align="center">
-  <a href="https://www.repostatus.org/#active"><img src="https://www.repostatus.org/badges/latest/active.svg" alt="Project Status: Active"></a>
-  <a href="https://pypi.org/project/solweig-gpu/"><img src="https://badge.fury.io/py/solweig-gpu.svg" alt="PyPI version"></a>
-  <a href="https://doi.org/10.5281/zenodo.17048978"><img src="https://zenodo.org/badge/DOI/10.5281/zenodo.17048978.svg" alt="DOI"></a>
-  <!-- <a href="https://anaconda.org/conda-forge/solweig-gpu"><img src="https://anaconda.org/conda-forge/solweig-gpu/badges/version.svg" alt="Conda version"></a> -->
-  <a href="https://www.gnu.org/licenses/gpl-3.0"><img src="https://img.shields.io/badge/License-GPLv3+-blue.svg" alt="License: GPL v3 or later"></a>
-  <a href="https://pepy.tech/project/solweig-gpu"><img src="https://pepy.tech/badge/solweig-gpu" alt="PyPI Downloads"></a>
-  <a href="https://joss.theoj.org/papers/27faa2bf5f6058d981df8b565f8e9a34"><img src="https://joss.theoj.org/papers/27faa2bf5f6058d981df8b565f8e9a34/status.svg"></a>
-</p>
-**SOLWEIG-GPU** is a powerful Python package for GPU-accelerated urban thermal comfort modeling. It enables high-resolution microclimate analysis by computing key variables such as Sky View Factor (SVF), Mean Radiant Temperature (Tmrt), and the Universal Thermal Climate Index (UTCI).
-
-## Overview
-
-SOLWEIG-GPU is designed to provide researchers, urban planners, and climate scientists with a fast and efficient tool for analyzing urban thermal environments. The package leverages GPU acceleration through PyTorch to process large urban areas quickly, making it practical for city-scale thermal comfort assessments.
-
-The model was originally developed by Dr. Fredrik Lindberg's group and has been extended with GPU acceleration capabilities. This implementation maintains compatibility with the original SOLWEIG model while providing significant performance improvements through parallel computing.
-
-## Key Features
-
-**GPU Acceleration** — The package utilizes PyTorch for high-performance computation on NVIDIA GPUs, with an automatic fallback to the CPU when a GPU is not available.
-
-**Comprehensive Modeling** — Calculates SVF, Tmrt, UTCI, and detailed shortwave and longwave radiation fluxes for complete thermal environment characterization.
-
-**Flexible Data Input** — Supports multiple meteorological data sources, including custom text files (UMEP format), ERA5 reanalysis data, and WRF model outputs.
-
-**Intelligent Tiling** — Automatically divides large study areas into manageable tiles with configurable overlap to handle memory constraints and ensure seamless results.
-
-**Multiple Interfaces** — Offers both a command-line interface for scripting and automation, and a graphical user interface for interactive use.
-
-**Parallel Processing** — CPU-based computations (wall heights and aspects) are parallelized across multiple cores for optimal performance.
-
-## Quick Links
-
-<table>
-  <tr>
-    <td>
-      <a href="./installation.md"><strong>Installation</strong></a><br/>
[…]
