### 2025-10-16 a91f6b1
Change Python version in installation guide
Updated Python version in installation instructions from 3.11 to 3.10.
