### 2025-10-16 099457c
Change documentation link in README
Updated the documentation link to point to Solweig-GPU.
