### 2025-10-16 eb3b363
Fix typo in Solweig-GPU documentation link
