### 2025-10-16 3c5a5e1
Revise installation instructions and dependencies
Updated installation instructions to clarify methods and added additional dependencies for Conda installation.
