### 2025-10-16 9bf4920
Add cuDNN installation step for GPU support
Added installation instructions for cuDNN if GPU is available.
