+++ docs/installation.md
@@ -18,6 +18,7 @@ conda activate solweig
 # Install dependencies via conda
 conda install -c conda-forge gdal pytorch timezonefinder matplotlib sip 
 pip install PyQt5
+conda install -c conda-forge cudnn #If GPU is available
 pip install solweig-gpu
 
 
