### 2025-10-17 d307731
Enhance conda installation instructions
Updated conda installation instructions to include additional dependencies.
