### 2025-10-17 4dbfccd
Clean up installation instructions for solweig-gpu
Remove redundant pip install command for solweig-gpu.
