### 2025-10-17 786b383
Delete docs/README.md
