### 2025-10-17 dac0a24
Add GitHub stars badge to index.rst
Added GitHub stars badge to documentation.
