### 2025-10-17 c9a15c3
Add documentation status badge to README
