### 2025-10-17 33e8159
made notebooks render in the docs
