### 2025-10-17 994e587
made notebooks render in the docs
