+++ docs/requirements.txt
@@ -4,3 +4,5 @@ sphinx-rtd-theme>=1.0
 myst-parser>=0.18
 sphinx-autodoc-typehints>=1.19
 linkify-it-py>=2.0.0
+nbsphinx>=0.9.0
+ipykernel>=6.0.0
