### 2025-10-17 e5833d2
Update .readthedocs.yml
