### 2025-10-17 6b61bd7
fix jupyternotebook
