### 2025-10-17 0b4bd54
Update notebooks.rst by removing introductory note
Removed introductory note about viewing notebooks.
