+++ .github/workflows/joss.yml
@@ -1,0 +1,33 @@
+name: Compile JOSS Paper
+
+on:
+  push:
+    branches:
+      - joss
+    paths:
+      - 'paper/**'
+      - '.github/workflows/joss.yml'
+  workflow_dispatch:
+
+jobs:
+  build-paper:
+    runs-on: ubuntu-latest
+
+    steps:
+      - name: Checkout repository
+        uses: actions/checkout@v3
+
+      - name: Set up Pandoc and LaTeX
+        run: |
+          sudo apt-get update
+          sudo apt-get install -y pandoc texlive-xetex
+
+      - name: Compile paper
+        working-directory: paper
+        run: make
+
+      - name: Upload compiled PDF
+        uses: actions/upload-artifact@v4
+        with:
+          name: joss-paper
+          path: paper/paper.pdf
