### 2025-10-17 bf0ddeb
Update Solweig-GPU documentation link in README
