### 2025-10-17 cf3b813
Bump version from 1.2.15 to 1.2.16
