+++ solweig_gpu/__init__.py
@@ -11,5 +11,5 @@
 #MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
 #GNU General Public License for more details.
 
-__version__ = "1.2.16"
+__version__ = "1.2.17"
 from .solweig_gpu import thermal_comfort
