+++ setup.py
@@ -47,7 +47,7 @@ setup(
     ],
     license="GPL-3.0-only",
     classifiers=[
-        "License :: OSI Approved :: GNU General Public License v3",
+        "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
         "Programming Language :: Python :: 3",
         "Operating System :: POSIX :: Linux",
         "Intended Audience :: Science/Research",
