### 2025-10-17 efcee76
Bump version from 1.2.17 to 1.2.18
