### 2025-10-17 12c8ad6
Bump version to 1.2.18
