+++ DOCS_BUILD_SUCCESS.md
@@ -1,288 +1,0 @@
-# ✅ Documentation Build - SUCCESS!
-
-## Status: READY FOR READ THE DOCS 🎉
-
-Your Sphinx documentation has been successfully configured and builds without errors!
-
----
-
-## ✅ What's Working
-
-### Local Build Test
-```bash
-cd docs && make html
-```
-**Result:** ✅ Build succeeded with all API modules documented!
-
-**Output:**
-- 81 functions documented across 9 modules
-- All docstrings successfully extracted
-- API reference auto-generated
-- Search index created
-- HTML pages built
-
-### Modules Successfully Documented:
-- ✅ `solweig_gpu.solweig_gpu` (2 functions)
-- ✅ `solweig_gpu.preprocessor` (16 functions)
-- ✅ `solweig_gpu.solweig` (24 functions)
-- ✅ `solweig_gpu.sun_position` (17 functions)
-- ✅ `solweig_gpu.shadow` (5 functions)
-- ✅ `solweig_gpu.calculate_utci` (2 functions)
-- ✅ `solweig_gpu.utci_process` (6 functions)
-- ✅ `solweig_gpu.walls_aspect` (6 functions)
-- ✅ `solweig_gpu.cli` (2 functions)
-- ✅ `solweig_gpu.Tgmaps_v1` (1 function)
-
----
-
-## 📁 Files Created/Modified
-
-### Configuration Files
-- ✅ `.readthedocs.yml` - Read the Docs configuration (fixed)
-- ✅ `docs/conf.py` - Sphinx configuration with mocking
-- ✅ `docs/requirements.txt` - Sphinx dependencies
-- ✅ `docs/Makefile` - Build commands
-
-### Documentation Content
-- ✅ `docs/index.rst` - Main documentation index
-- ✅ `docs/api.rst` - API reference structure
-- ✅ `docs/installation.md` - Installation guide
-- ✅ `docs/quickstart.md` - Quick start tutorial
-- ✅ `docs/testing.md` - Testing documentation
-- ✅ `docs/README.md` - Documentation maintenance guide
-
-### Required Directories
-- ✅ `docs/_static/` - Static files (images, CSS)
-- ✅ `docs/_templates/` - Custom templates
-
----
-
-## 🔧 Issues Fixed
-
-### 1. Deprecated Configuration ✅
-**Before:**
-```yaml
-python:
-  version: "3.11"  # ❌ Deprecated key
-```
-
-**After:**
-```yaml
-build:
-  tools:
-    python: "3.11"  # ✅ Correct format
-```
-
-### 2. GDAL Installation Failure ✅
-**Problem:** Read the Docs doesn't have GDAL system libraries
-
-**Solution:** Mock all dependencies instead of installing package
-```python
-autodoc_mock_imports = [
-    'torch', 'gdal', 'osgeo', 'scipy', 'numpy',
-    'pandas', 'xarray', 'shapely', 'netCDF4',
-    'PyQt5', 'timezonefinder', 'pytz'
-]
-```
-
-### 3. Missing Directories ✅
-**Problem:** Sphinx expects `_static` and `_templates` directories
-
-**Solution:** Created directories with `.gitkeep` files
-
----
-
-## ⚠️ Minor Warnings (Non-Critical)
-
-The build has 30 warnings about missing cross-references:
-```
-WARNING: 'myst' cross-reference target not found: 'guide/input_data.md'
-WARNING: 'myst' cross-reference target not found: '../CONTRIBUTING.md'
-```
-
-**These are OK!** They're just broken links in documentation - the build still succeeds and all API docs work perfectly.
-
-**Optional Fix:** Create the missing guide pages or update the links.
-
----
-
[…]
