### 2025-10-17 ca38579
Adjust end_time and enhance ERA-5 data explanation
Updated end_time in configuration and clarified ERA-5 data handling in README.
