### 2025-10-17 b1579cd
Update README with ERA-5 datetime processing details
Clarified automatic datetime setting for ERA-5 data processing.
