### 2025-10-17 b7f66fc
Update configuration.md
