### 2025-10-17 3dc2bf5
Revise Meteorological Data Configuration sections
Updated section titles for clarity and added context about sample data usage.
