### 2025-10-17 fe2fd47
Revise examples title and end time in documentation
Updated the title and modified the end time in Example 1.
