### 2025-10-17 b1da7dd
Add CDS API installation instructions to examples
Added installation instructions for CDS API to download ERA5 data.
