### 2025-10-17 f0c1cf3
Update SOLWEIG-GPU description with new features
Expanded the description of SOLWEIG-GPU to include additional calculations.
