### 2025-10-17 6be6504
Update quickstart guide with sample data link
Added a section for sample data availability in Zenodo.
