### 2025-10-17 1d865a3
Change clone URL in installation instructions
Updated the repository URL for cloning.
