### 2025-10-17 25573ee
Adjust start and end times for data examples
Updated time parameters for ERA5 and WRF examples to start at 06:00 UTC instead of 00:00 UTC.
