### 2025-10-17 5a9b5fc
Fix indentation and formatting in index.rst
