### 2025-10-17 311ad9c
Modify end_time and add ERA5 data download guide
Updated end_time for WRF data and added ERA5 download instructions.
