### 2025-10-17 5bfb0a9
Fix typos and improve clarity in configuration.md
