### 2025-10-17 94fefab
Fix formatting and improve clarity in configuration.md
