### 2025-10-17 d6dc286
Improve memory allocation guidance and cdsapi setup
Clarified memory allocation advice for tile size adjustments and updated the setup instructions for cdsapi.
