### 2025-10-17 71916bd
Add sample data availability note in examples.md
Added a note about sample data availability in Zenodo.
