+++ docs/examples.md
@@ -2,6 +2,8 @@
 
 This section provides a collection of examples to demonstrate how to use SOLWEIG-GPU for different scenarios.
 
+Sample data is available in [Zenodo](https://zenodo.org/records/17048978)
+
 ## Example 1: Using WRF Data
 
 This example shows how to run a simulation using meteorological data from WRF output files.
