### 2025-10-17 f356755
Fix formatting and grammar in paper.md
