+++ paper/paper.md
@@ -53,33 +53,33 @@ The original SOLWEIG model [@lindberg2008solweig] was developed to calculate T<s
 # Functionality 
 SOLWEIG-GPU requires the following inputs: (i) Building digital surface model (DSM) that includes both buildings and terrain height (*m*), (ii) Digital elevation model (DEM) that is the bare ground elevation (*m*), (iii). Tree or vegetation DSM that only represents the vegetation height (*m*) (iv). UMEP [@lindberg2018umep] style ground cover (optional), and (v) meteorological forcing. Allowed land cover types are asphalt or paved, buildings, grass, bare soil, and water but users can add their own land cover types provided they know the thermal properties for the land surface (e.g., albedo and emissivity). All the input datasets must be of the same spatial resolution, projection, and spatial extent. The recommended projection is the Universal Transverse Mercator (UTM). For example, EPSG: 32614. Necessary meteorological variables for SOLWEIG are: (i) 2-meter air temperature (*℃*), (ii) relative humidity (*%*), (iii) barometric pressure (*kPa*), (iv) downwelling shortwave radiation (*W/m<sup>2</sup>*). Additionally, near-surface wind speed (*m/s*) is required for UTCI computation and longwave radiation is estimated using 2-meter air temperature, relative humidity and pressure. 
 
-![Different steps involved in calculation of `thermal comfort` using SOLWEIG-GPU: CPU and GPU based calculations in SOLWEIG-GPU are shown.](figures/figure1.png)
+![Different steps involved in calculation of `thermal comfort` using SOLWEIG-GPU: CPU and GPU-based calculations in SOLWEIG-GPU are shown.](figures/figure1.png)
 
-Figure 1 shows the workflow of SOLWEIG-GPU and detailed description of functionalities are as follows:
+Figure 1 shows the workflow of SOLWEIG-GPU, and a detailed description of functionalities is as follows:
   1. SOLWEIG-GPU can divide a larger geographical domain into smaller tiles and create separate meteorological forcing for each of the tiles. 
   2. If **torch.cuda.is_available()** returns **‘True’**, the simulations will be performed on a GPU.
   3. The calculations of wall height and aspect (wall directional orientation) are faster on CPU as they use an *i* and *j* indices to loop through the input rasters. Thus, we have parallelized this operation and calculations are performed on multiple CPUs at once.
-  4. SVF, T<sub>MRT</sub>, UTCI and ground shading are computed on the GPU, if available. Additionally, the model can output shortwave and longwave radiations in both upward and downward directions. The users can select the required output variables from SOLWEIG-GPU but UTCI is outputted by default. Note that calculation of UTCI uses wind speed from the meteorological forcing but more advanced methods to calculate wind speeds in urban areas are available [@bernard2023urock]. On clear, hot days T<sub>MRT</sub> is a dominant driver, yet wind can still change UTCI by several kelvin, so relying on grid-averaged wind may not always have a minimal impact.
+  4. SVF, T<sub>MRT</sub>, UTCI and ground shading are computed on the GPU, if available. Additionally, the model can output shortwave and longwave radiations in both upward and downward directions. The users can select the required output variables from SOLWEIG-GPU, but UTCI is output by default. Note that calculation of UTCI uses wind speed from the meteorological forcing but more advanced methods to calculate wind speeds in urban areas are available [@bernard2023urock]. On clear, hot days T<sub>MRT</sub> is a dominant driver, yet wind can still change UTCI by several kelvin, so relying on grid-averaged wind may not always have a minimal impact.
   5. Ground cover classes can be optionally provided and they are used to set-up the grid for the surface thermal properties [@lindberg2016ground].
[…]
