### 2025-10-20 28a26fd
add TILESLICING_SOLWEIG, Andrea Zonato, To merge
