+++ TILESLICING_SOLWEIG/solweig_gpu/solweig_gpu_gui.py
@@ -1,0 +1,425 @@
+import sys
+import os
+from PyQt5.QtWidgets import (
+    QApplication, QWidget, QVBoxLayout, QLabel, QPushButton, QLineEdit, QFileDialog, QCheckBox,
+    QDateEdit, QSpinBox, QComboBox, QProgressBar, QTextEdit, QGroupBox, QFormLayout,
+    QHBoxLayout, QFrame, QToolButton, QStyle, QMessageBox, QScrollArea, QDesktopWidget
+)
+from PyQt5.QtCore import QThread, pyqtSignal, Qt
+from solweig_gpu import thermal_comfort
+
+
+class EmittingStream:
+    def __init__(self, signal):
+        self.signal = signal
+
+    def write(self, text):
+        self.signal.emit(str(text))
+
+    def flush(self):
+        pass
+
+
+class SOLWEIGWorker(QThread):
+    log_signal = pyqtSignal(str)
+    finished = pyqtSignal()
+
+    def __init__(self, params):
+        super().__init__()
+        self.params = params
+
+    def run(self):
+        sys.stdout = EmittingStream(self.log_signal)
+        try:
+            thermal_comfort(**self.params)
+        except Exception as e:
+            self.log_signal.emit(f"\nError: {str(e)}")
+        finally:
+            self.finished.emit()
+
+
+class SOLWEIGApp(QWidget):
+    def __init__(self):
+        super().__init__()
+        self.setWindowTitle("SOLWEIG GPU")
+
+        screen = QDesktopWidget().screenGeometry()
+        width = int(screen.width() * 0.85)
+        height = int(screen.height() * 0.85)
+        self.resize(width, height)
+
+        self.setStyleSheet(self.load_styles())
+
+        scroll = QScrollArea()
+        scroll.setWidgetResizable(True)
+        container = QWidget()
+        self.main_layout = QVBoxLayout(container)
+        scroll.setWidget(container)
+
+        outer_layout = QVBoxLayout(self)
+        outer_layout.addWidget(scroll)
+        self.setLayout(outer_layout)
+
+        self._build_widgets()
+
+
+    def select_path(self, target_widget, file=True):
+        if file:
+            path = QFileDialog.getOpenFileName()[0]
+        else:
+            path = QFileDialog.getExistingDirectory()
+        if path:
+            target_widget.setText(path)
+
+
+    def _build_widgets(self):
+        self.log_output = QTextEdit()
+        self.log_output.setReadOnly(True)
+        self.log_output.setStyleSheet("background-color: #1e1e1e; color: #00ff00; font-family: monospace;")
+
+        self.progress = QProgressBar()
+        self.progress.setMinimum(0)
+        self.progress.setMaximum(100)
+        self.progress.setValue(0)
+
+        self.run_button = QPushButton("Run SOLWEIG")
+        self.run_button.setCursor(Qt.PointingHandCursor)
+        self.run_button.setStyleSheet("""
+            QPushButton {
+                background-color: #00c853;
+                color: white;
+                border-radius: 6px;
+                font-size: 16px;
+                font-weight: bold;
+                padding: 10px 20px;
+                margin: 10px auto;
+            }
+            QPushButton:hover {
+                background-color: #00e676;
+            }
+            QPushButton:disabled {
+                background-color: #555;
+                color: #999;
+            }
+        """)
+        self.run_button.clicked.connect(self.run_solweig)
+
+        header = QLabel("SOLWEIG GPU")
+        header.setAlignment(Qt.AlignCenter)
+        header.setStyleSheet("font-size: 24px; font-weight: bold; margin: 10px 0; color: #ffffff;")
+        self.main_layout.addWidget(header)
+
[…]
