+++ TILESLICING_SOLWEIG/solweig_gpu/__init__.py
@@ -1,0 +1,5 @@
+"""SOLWEIG GPU refactored interface."""
+
+from .runtime import thermal_comfort
+
+__all__ = ["thermal_comfort"]
