+++ TILESLICING_SOLWEIG/create_inputs_solweig.py
@@ -1,0 +1,2310 @@
+
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+
+print("Developed by Andrea Zonato, andrea.zonato@cimafoundation.org")
+
+"""Build input datasets for SOLWEIG.
+
+Author: Andrea Zonato
+
+What this script does
+- Given a target location (lat, lon), it computes geographic bounding boxes
+  and a UTM working CRS, then prepares a common analysis grid for SOLWEIG.
+- It collects static inputs: landcover (ESA WorldCover), canopy height/trees
+  (multiple Earth Engine sources), and a DEM (TINITALY if in Italy, otherwise
+  MERIT/DEM via Earth Engine). It also derives urban layers from OSM (vegetation,
+  impervious, water) and buildings from a public GBA WFS (LOD1), with an OSM fallback.
+- It rasterizes vector layers (vegetation, water, buildings, urban surfaces),
+  resamples everything to the common grid, applies stacking logic (water >
+  vegetation > urban), and builds a Building DSM = DEM + building heights.
+- It downloads and prepares meteorological forcing. If the area is in Italy,
+  it uses the DDS downscaled dataset; elsewhere it samples ERA5 via Earth Engine.
+  In both cases, it standardizes names/dimensions and derives relative
+  humidity and surface altitude to save a single, consistent NetCDF file.
+
+"""
+
+import argparse, math, shutil, subprocess, sys, time, io
+from pathlib import Path
+
+import importlib, importlib.util
+from dataclasses import dataclass
+from typing import Any, List
+
+# Bootstrap: ensure third-party packages are installed before importing
+def _ensure(mod, pip_name=None, required=True):
+    """Ensure module is installable; avoid importing until actually needed."""
+    if importlib.util.find_spec(mod) is None:
+        pkg = pip_name or mod
+        try:
+            subprocess.check_call([sys.executable, "-m", "conda", "install", pkg])
+        except Exception as e:
+            if required:
+                raise
+            else:
+                print(f"[deps] Optional package '{pkg}' not installed: {e}")
+
+for _mod, _pip in [
+    ("ee", "earthengine-api"), ("geemap", None), ("geopandas", None), ("numpy", None),
+    ("pandas", None), ("rasterio", None), ("requests", None), ("osmnx", None),
+    ("scipy", None), ("geopy", None), ("pyproj", None), ("shapely", None), ("xarray", None),
+    ("cftime", None), ("netCDF4", None), ("ddsapi", None)
+]:
+    _ensure(_mod, _pip)
+
+import ee
+import geemap
+import geopandas as gpd
+import numpy as np
+import pandas as pd
+import rasterio
+import requests
+import osmnx as ox
+from geopy.geocoders import Nominatim
+from pyproj import Transformer
+from rasterio.features import rasterize
+from rasterio.merge import merge
+from rasterio.transform import from_origin
+from rasterio.warp import Resampling, reproject
+from shapely.geometry import box
+import xarray as xr
[…]
