+++ TILESLICING_SOLWEIG/solweig_gpu/solweig_gpu_multigpu.py
@@ -1,0 +1,7 @@
+"""Compatibility layer retaining old import paths."""
+
+from __future__ import annotations
+
+from .runtime import thermal_comfort
+
+__all__ = ["thermal_comfort"]
