### 2025-10-22 bb2d83c
Revise description of SOLWEIG-GPU model
Clarified the purpose and functionality of the SOLWEIG-GPU model, emphasizing its GPU acceleration for urban thermal comfort simulations.
