### 2025-10-22 73e7fcb
Revise introduction in documentation
Removed introductory sentence about GPU-accelerated SOLWEIG model.
