### 2025-10-22 29b0fcd
Enhance documentation for SOLWEIG-GPU framework
Added details about the SOLWEIG-GPU framework and its features.
