### 2025-10-22 a34a49f
Refactor SOLWEIG-GPU documentation introduction
Reorganized and clarified the introduction of the SOLWEIG-GPU documentation for better readability.
