### 2025-10-22 977ee2e
Clarify GPU SVF tool details in documentation
Updated the documentation to clarify the existence of a GPU SVF computation tool and its integration with Python and C.
