### 2025-10-22 e1974bb
Fix formatting of GPU SVF tool reference in docs
