### 2025-10-22 db7a9f4
Fix formatting of GPU tool link in documentation
Updated the documentation to format the GPU tool link correctly.
