### 2025-10-22 23a596b
Enhance documentation with SOLWEIG model link
Updated the documentation to include a link to the original SOLWEIG model and clarified the GPU acceleration benefits for city-scale thermal comfort estimation.
