### 2025-10-22 d119fe7
Update accuracy description for SOLWEIG algorithms
