### 2025-10-29 dedea7a
Update README with solweig-gpu upgrade instructions
Added instructions to upgrade solweig-gpu if older versions are installed.
