### 2025-10-29 f22c4af
Update installation instructions for solweig-gpu
Added instructions to upgrade solweig-gpu if older versions are installed.
