### 2025-10-31 7f34ea3
Create CONTRIBUTING.md for developer guidelines
Added a developer guide for contributing to the SOLWEIG-GPU project, including architecture overview and contribution steps.
