### 2025-11-19 149b2bc
Shorten the paper
Updated the paper to improve clarity and conciseness in the description of SOLWEIG-GPU inputs, outputs, and funding sources.
