+++ paper/paper.md
@@ -51,7 +51,7 @@ The original SOLWEIG model [@lindberg2008solweig] was developed to calculate T<s
 
 
 # Functionality 
-SOLWEIG-GPU requires the following inputs: (i) Building digital surface model (DSM) that includes both buildings and terrain height (*m*), (ii) Digital elevation model (DEM) that is the bare ground elevation (*m*), (iii). Tree or vegetation DSM that only represents the vegetation height (*m*) (iv). UMEP [@lindberg2018umep] style ground cover (optional), and (v) meteorological forcing. Allowed land cover types are asphalt or paved, buildings, grass, bare soil, and water but users can add their own land cover types provided they know the thermal properties for the land surface (e.g., albedo and emissivity). All the input datasets must be of the same spatial resolution, projection, and spatial extent. The recommended projection is the Universal Transverse Mercator (UTM). For example, EPSG: 32614. Necessary meteorological variables for SOLWEIG are: (i) 2-meter air temperature (*℃*), (ii) relative humidity (*%*), (iii) barometric pressure (*kPa*), (iv) downwelling shortwave radiation (*W/m<sup>2</sup>*). Additionally, near-surface wind speed (*m/s*) is required for UTCI computation and longwave radiation is estimated using 2-meter air temperature, relative humidity and pressure. 
+SOLWEIG-GPU requires the following inputs: (i) Building digital surface model (DSM) that includes both buildings and terrain height (*m*), (ii) Digital elevation model (DEM) that is the bare ground elevation (*m*), (iii). Tree or vegetation DSM that only represents the vegetation height (*m*) (iv). UMEP [@lindberg2018umep] style ground cover (optional), and (v) meteorological forcing. Allowed land cover types are asphalt or paved, buildings, grass, bare soil, and water but users can add their own land cover types provided they know the thermal properties for the land surface (e.g., albedo and emissivity). All the input datasets must be of the same spatial resolution, projection, and spatial extent. The recommended projection is the Universal Transverse Mercator (UTM). For example, EPSG: 32614 (for Austin, TX). Necessary meteorological variables for SOLWEIG are: (i) 2-meter air temperature (*℃*), (ii) relative humidity (*%*), (iii) barometric pressure (*kPa*), (iv) downwelling shortwave radiation (*W/m<sup>2</sup>*). Additionally, near-surface wind speed (*m/s*) is required for UTCI computation. Longwave radiation is estimated using 2-meter air temperature, relative humidity and pressure. 
 
 ![Different steps involved in calculation of `thermal comfort` using SOLWEIG-GPU: CPU and GPU-based calculations in SOLWEIG-GPU are shown.](figures/figure1.png)
 
@@ -59,7 +59,7 @@ Figure 1 shows the workflow of SOLWEIG-GPU, and a detailed description of functi
   1. SOLWEIG-GPU can divide a larger geographical domain into smaller tiles and create separate meteorological forcing for each of the tiles. 
   2. If **torch.cuda.is_available()** returns **‘True’**, the simulations will be performed on a GPU.
   3. The calculations of wall height and aspect (wall directional orientation) are faster on CPU as they use an *i* and *j* indices to loop through the input rasters. Thus, we have parallelized this operation and calculations are performed on multiple CPUs at once.
[…]
