### 2025-11-19 49962bf
Shorten the paper
Updated usage instructions for SOLWEIG-GPU, removing detailed example and providing links to documentation.
