### 2025-11-19 a10eb24
Shorten the paper
Updated the description of SOLWEIG-GPU to clarify data sources and inputs. Added examples and improved details on model usage and comparisons with CPU-based simulations.
