+++ paper/paper.md
@@ -43,7 +43,7 @@ bibliography: paper.bib
 
 # Summary
 
-We present Solar and LongWave Environmental Irradiance Geometry-Graphics Processing Unit (SOLWEIG-GPU) as a Python package that provides a single-line code, command-line interface (CLI) and a graphical user interface (GUI) for executing the SOLWEIG model version 2022a on GPUs. The package facilitates GPU parallelized human thermal comfort modeling at meter-scale resolution across city-scale domains by computing key variables such as sky-view factor (SVF), mean radiant temperature (T<sub>MRT</sub>), ground shading, and the universal thermal climate index (UTCI). Sample data and model instructions can be found in the repository.
+We present Solar and LongWave Environmental Irradiance Geometry-Graphics Processing Unit (SOLWEIG-GPU) as a Python package that provides a single-line code, command-line interface (CLI) and a graphical user interface (GUI) for executing the SOLWEIG model version 2022a on GPUs. The package facilitates GPU parallelized human thermal comfort modeling at meter-scale resolution across city-scale domains by computing key variables such as sky-view factor (SVF), mean radiant temperature (T<sub>MRT</sub>), ground shading, and the universal thermal climate index (UTCI). Sample data for running the model can be found on Zenodo and quick start guide in the documentation.
 
 # Statement of need
 
@@ -51,7 +51,7 @@ The original SOLWEIG model [@lindberg2008solweig] was developed to calculate T<s
 
 
 # Functionality 
-SOLWEIG-GPU requires the following inputs: (i) Building digital surface model (DSM) that includes both buildings and terrain height (*m*), (ii) Digital elevation model (DEM) that is the bare ground elevation (*m*), (iii). Tree or vegetation DSM that only represents the vegetation height (*m*) (iv). UMEP [@lindberg2018umep] style ground cover (optional), and (v) meteorological forcing. Allowed land cover types are asphalt or paved, buildings, grass, bare soil, and water but users can add their own land cover types provided they know the thermal properties for the land surface (e.g., albedo and emissivity). All the input datasets must be of the same spatial resolution, projection, and spatial extent. The recommended projection is the Universal Transverse Mercator (UTM). For example, EPSG: 32614 (for Austin, TX). Necessary meteorological variables for SOLWEIG are: (i) 2-meter air temperature (*℃*), (ii) relative humidity (*%*), (iii) barometric pressure (*kPa*), (iv) downwelling shortwave radiation (*W/m<sup>2</sup>*). Additionally, near-surface wind speed (*m/s*) is required for UTCI computation. Longwave radiation is estimated using 2-meter air temperature, relative humidity and pressure. 
[…]
