### 2025-11-19 2cbd2e2
Refine language and clarity in paper.md
Corrected grammatical errors and improved clarity in the paper's text.
