+++ paper/paper.md
@@ -43,37 +43,36 @@ bibliography: paper.bib
 
 # Summary
 
-We present Solar and LongWave Environmental Irradiance Geometry-Graphics Processing Unit (SOLWEIG-GPU) as a Python package that provides a single-line code, command-line interface (CLI) and a graphical user interface (GUI) for executing the SOLWEIG model version 2022a on GPUs. The package facilitates GPU parallelized human thermal comfort modeling at meter-scale resolution across city-scale domains by computing key variables such as sky-view factor (SVF), mean radiant temperature (T<sub>MRT</sub>), ground shading, and the universal thermal climate index (UTCI). Sample data for running the model can be found on Zenodo and quick start guide in the documentation.
+We present Solar and LongWave Environmental Irradiance Geometry-Graphics Processing Unit (SOLWEIG-GPU) as a Python package that provides a single-line code, command-line interface (CLI), and a graphical user interface (GUI) for executing the SOLWEIG model version 2022a on GPUs. The package facilitates GPU-parallelized human thermal comfort modeling at meter-scale resolution across city-scale domains by computing key variables, such as sky-view factor (SVF), mean radiant temperature (T<sub>MRT</sub>), ground shading, and the universal thermal climate index (UTCI). Sample data for running the model can be found on Zenodo, and a quick-start guide is included in the documentation.
 
 # Statement of need
 
-The original SOLWEIG model [@lindberg2008solweig] was developed to calculate T<sub>MRT</sub> over small geographical areas in cities. At this spatial scale, based on the time required for computation, the model could be run on CPUs [@kamath2023human]. However, for city-scale thermal comfort estimation, the model can be accelerated using a GPU. Specifically, the calculation of the SVF (which is the most time-consuming step), T<sub>MRT</sub> and UTCI can be processed on a GPU. Currently, there is a tool that computes SVF on GPU but it uses Python to read the inputs and write the output rasters, while the SVF calculation on GPU is done by interfacing with a C-language code [@li2021gpu]. Thus, there was a need for a Python-based end-to-end framework to run SOLWEIG on a GPU. Our framework is implemented in PyTorch, which enables automatic selection of the GPU when available. 
-
+The original SOLWEIG model [@lindberg2008solweig] was developed to calculate T<sub>MRT</sub> over small geographical areas in cities. At a small spatial scale, based on the time required for computation, the model could be run on CPUs [@kamath2023human]. However, for city-scale thermal comfort estimation, the model can be accelerated using a GPU. Specifically, the calculation of the SVF (the most time-consuming step), T<sub>MRT</sub>, and UTCI can be processed on a GPU. Currently, there is a tool that computes SVF on GPU, but it uses Python to read the inputs and write the output rasters, while the SVF calculation on GPU is done by interfacing with a C-language code [@li2021gpu]. Thus, there was a need for a Python-based end-to-end framework to run SOLWEIG on a GPU. Our framework is implemented in PyTorch, which enables automatic selection of the GPU when available. 
 
 # Functionality 
[…]
