### 2025-12-02 #8944 @adithirgis
> > 👋 [@lukasbeuster](https://github.com/lukasbeuster) - Would you be willing to review this submission to JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
>
> Hi Adithi, I'd be happy to review this submission. Since the authors tagged me, I've had a student pick this up and she's using it for her thesis. Unless this disqualifies me I can add some comments from the actual use of this package.

Hello @lukasbeuster I think it should be completely fine to review it unless the student or you had any collaboration with the authors which I assume is not the case.

Thank you! I will be adding you as a reviewer.
