### 2025-12-02 #9535 @editorialbot
[REVIEW]: SOLWEIG-GPU: GPU-Accelerated Thermal Comfort Modeling Framework for Urban Digital Twins
**Submitting author:** @nvnsudharsan (<a href="http://orcid.org/0000-0002-1328-110X">Naveen Sudharsan</a>)
**Repository:** https://github.com/nvnsudharsan/SOLWEIG-GPU
**Branch with paper.md** (empty if default branch): joss
**Version:** v1.2.21
