### 2025-12-09 nvnsudharsan/SOLWEIG-GPU#78 @ruedi99ms
Conda Usage
The authors should mention the recommend usage of Conda in the readme file, this is only mentioned in the full documentation ("# Using conda (recommended)"). Currently, the readme file starts with the usage of the conda command - unexperienced users might face problems here, as the command might pop up as unknown in a terminal. https://github.com/openjournals/joss-reviews/issues/9535
