### 2025-12-09 nvnsudharsan/SOLWEIG-GPU#79 @ruedi99ms
Paper: Functionality
In the functionality section of the paper under point 2, it is not stated, what will happen when cuda.is_available() returns false

I suppose, the calculations will then not be performed on a GPU (but CPU) -> this is not reflected in the figure

https://github.com/openjournals/joss-reviews/issues/9535
