### 2025-12-09 e8e533f
Recommend using conda for environment setup
Added conda environment setup instructions to README.
