### 2025-12-10 b2233c9
Update installation instructions in README
Added a reference to the documentation for conda environment setup.
