### 2025-12-10 7778dd8
Update installation documentation link in README
