+++ README.md
@@ -79,7 +79,7 @@ For detailed documentation, see [Solweig-GPU Documentation](https://solweig-gpu.
 
 ## Installation
 
-We recommend using conda environment (please see [documentation](./docs/developer_guide.md))
+We recommend using conda environment (please see [documentation](./docs/installation.md))
 
 ```bash
 conda create -n solweig python=3.10
