### 2025-12-10 63dfdb0
Fix typos and enhance documentation clarity
Corrected typos and improved clarity in the documentation for SOLWEIG-GPU, including usage instructions and input requirements.
