+++ paper/paper.md
@@ -47,28 +47,28 @@ We present Solar and LongWave Environmental Irradiance Geometry-Graphics Process
 
 # Statement of need
 
-The original SOLWEIG model [@lindberg2008solweig] was developed to calculate T<sub>MRT</sub> over small geographical areas in cities. At this spatial scale, based on the time required for computation, the model could be run on CPUs [@kamath2023human]. However, for city-scale thermal comfort estimation, the model can be accelerated using a GPU. Specifically, the calculation of the SVF (which is the most time-consuming step), T<sub>MRT</sub> and UTCI can be processed on a GPU. Currently, there is a tool that computes SVF on GPU but it uses Python to read the inputs and write the output rasters, while the SVF calculation on GPU is done by interfacing with a C-language code [@li2021gpu]. Thus, there was a need for a Python-based end-to-end framework to run SOLWEIG on a GPU. Our framework is implemented in PyTorch, which enables automatic selection of the GPU when available. 
+The original SOLWEIG model [@lindberg2008solweig] was developed to calculate T<sub>MRT</sub> over small geographical areas in cities. At this spatial scale, based on the time required for computation, the model could be run on CPUs [@kamath2023human]. However, for city-scale thermal comfort estimation, the model can be accelerated using a GPU. Specifically, the calculation of the SVF (which is the most time-consuming step), T<sub>MRT</sub> and UTCI can be processed on a GPU. Currently, there is a tool that computes SVF on a GPU, but it uses Python to read the inputs and write the output rasters, while the SVF calculation on the GPU is done by interfacing with a C-language code [@li2021gpu]. Thus, there was a need for a Python-based end-to-end framework to run SOLWEIG on a GPU. Our framework is implemented in PyTorch, which enables automatic selection of the GPU when available. 
 
 
 # Functionality 
-SOLWEIG-GPU requires the following inputs: (i) Building digital surface model (DSM) that includes both buildings and terrain height (*m*), (ii) Digital elevation model (DEM) that is the bare ground elevation (*m*), (iii). Tree or vegetation DSM that only represents the vegetation height (*m*) (iv). UMEP [@lindberg2018umep] style ground cover (optional), and (v) meteorological forcing. Allowed land cover types are asphalt or paved, buildings, grass, bare soil, and water but users can add their own land cover types provided they know the radiative properties of the surfaces (e.g., albedo and emissivity). All the input datasets must be of the same spatial resolution, projection, and spatial extent. The recommended projection is the Universal Transverse Mercator (UTM). For example, EPSG: 32614 (for Austin, TX). Necessary meteorological variables for SOLWEIG are: (i) 2-meter air temperature (*℃*), (ii) relative humidity (*%*), (iii) barometric pressure (*kPa*), (iv) downwelling shortwave radiation (*W/m<sup>2</sup>*). Additionally, near-surface wind speed (*m/s*) is required for UTCI computation. Longwave radiation is estimated using 2-meter air temperature, relative humidity and pressure. 
[…]
