### 2025-12-10 nvnsudharsan/SOLWEIG-GPU#79 @nvnsudharsan
Changed it as below.
"If torch.cuda.is_available() returns ‘True’, the simulations will be performed on a GPU. Otherwise, computation defaults to the CPU."
