### 2025-12-11 nvnsudharsan/SOLWEIG-GPU#80 @ruedi99ms
Running CLI Test ERA5 Documentation Issue and Error
I have been trying to run the CLI version of your software. The documentation and the readme file vary:

- Documentation: Define start and end time via "--start_time" and "--end_time" -> These are not recognized arguments (s. [here](https://solweig-gpu.readthedocs.io/en/latest/quickstart.html#command-line-usage) )

- Readme: Define via "--start" and "--end" -> These are working, but lead to the following error when using the sample data provided: ValueError: No instantaneous timestamps within requested window.

I am using the sample data from Zenodo.

https://github.com/openjournals/joss-reviews/issues/9535
