### 2025-12-15 522acb4
Update command options in quickstart documentation
corrected the CLI options
