### 2025-12-18 nvnsudharsan/SOLWEIG-GPU#80 @ruedi99ms
Same problem for the regular run (via python script directly)

wrfout is detecting the timestamps (no error here)
