### 2025-12-18 nvnsudharsan/SOLWEIG-GPU#82 @ruedi99ms
Klyne API
Your software uses the Klyne API, this is not mentioned in the docs

As it seems as you are providing an API key in your __init__.py file, I am currently not able to verify the performance, as the API limit is reached

Also, in general, the undocumented API use can lead to issues (Klyne API error 400: Bad Request)

https://github.com/openjournals/joss-reviews/issues/9535
