### 2025-12-18 nvnsudharsan/SOLWEIG-GPU#83 @ruedi99ms
Wall Height and Aspect Error
Computing Wall Height and Aspect:   0%|                                                                                                                                                                                                             | 0/16 [00:09<?, ?it/s]
Traceback (most recent call last):
  File "c:\Users\admin\Downloads\SOLWEIG-GPU-main\thermal_test.py", line 3, in <module>
    thermal_comfort(
  File "c:\Users\admin\Downloads\SOLWEIG-GPU-main\solweig_gpu\solweig_gpu.py", line 49, in thermal_comfort
    run_parallel_processing(building_dsm_dir, walls_dir, aspect_dir)
  File "c:\Users\admin\Downloads\SOLWEIG-GPU-main\solweig_gpu\walls_aspect.py", line 157, in run_parallel_processing
    _ = future.result()
  File "C:\Users\admin\anaconda3\envs\solweig\lib\concurrent\futures\_base.py", line 451, in result
    return self.__get_result()
  File "C:\Users\admin\anaconda3\envs\solweig\lib\concurrent\futures\_base.py", line 403, in __get_result
    raise self._exception
concurrent.futures.process.BrokenProcessPool: A process in the process pool was terminated abruptly while the future was running or pending.

I am getting this error when trying to run the program with wrfout data in a standalone python script

https://github.com/openjournals/joss-reviews/issues/9535
