### 2025-12-18 nvnsudharsan/SOLWEIG-GPU#84 @ruedi99ms
WRFOUT Sample Data
The GUI is printing the following statement for the sample landcover data provided:

Attention! The land cover grid includes values (deciduous and/or conifer) not appropriate for the SOLWEIG-formatted land cover grid (should not include 3 or 4). Land cover under the vegetation is required. Setting the invalid landcover types to grass.

The sample data should either be correctly formatted or left out.

https://github.com/openjournals/joss-reviews/issues/9535
