### 2025-12-18 nvnsudharsan/SOLWEIG-GPU#85 @ruedi99ms
Output Folder
It would be nice, if the user can provide a file path for the output - so, that not everything is written into the data folder

https://github.com/openjournals/joss-reviews/issues/9535
