### 2025-12-18 #9535 @ruedi99ms
I am currently not able to fully complete the review, as there is an undocumented API issue and a usage limit for Klyne (I have raised an issue for that) - at least when using the api
