### 2026-01-15 4b0207d
Modify thermal_comfort command options
Updated the CLI usage
