### 2026-01-15 nvnsudharsan/SOLWEIG-GPU#80 @nvnsudharsan
The sample file included in Zenodo was mistakenly taken from 2000-08-13. This has been corrected to 2020-08-13 and will be updated in Zenodo soon.
The CLI documentation has also been updated to reflect the correct usage of the --start and --end arguments.
