### 2026-01-16 acdcf05
Refactor output paths and create preprocessing directory
Updated output directory structure and added preprocessing directory creation.
