### 2026-01-16 40aa865
Updated the output folder for pre processing outputs
After preprocessing, the input data will be saved in the processed_input folder inside the base path
