+++ solweig_gpu/preprocessor.py
@@ -131,7 +131,7 @@ def check_rasters(files):
 # =============================================================================
 # Function to tile a raster file into smaller chunks.
 # =============================================================================
-def create_tiles(infile, tilesize, overlap, tile_type):
+def create_tiles(infile, tilesize, overlap, tile_type, preprocess_dir):
     """
     Tile a raster file into smaller chunks.
 
@@ -140,6 +140,7 @@ def create_tiles(infile, tilesize, overlap, tile_type):
         tilesize (int): Size of each tile in pixels.
         overlap (int): Number of pixels to overlap between tiles.
         tile_type (str): Label to use for naming output tiles.
+        preprocess_dir (str): Directory to save tiles in the pre_processing_outputs folder.
 
     Raises:
         FileNotFoundError: If the input file is not found.
@@ -156,7 +157,7 @@ def create_tiles(infile, tilesize, overlap, tile_type):
     width = ds.RasterXSize
     height = ds.RasterYSize
 
-    out_folder = os.path.join(os.path.dirname(infile), tile_type)
+    out_folder = os.path.join(preprocess_dir, tile_type)
     if not os.path.exists(out_folder):
         os.makedirs(out_folder)
     else:
@@ -189,7 +190,7 @@ def create_tiles(infile, tilesize, overlap, tile_type):
         
 # It processes the data to compute temperature (in °C), surface pressure (in kPa), 
 # relative humidity (in %), wind speed (in m/s), shortwave and longwave radiation (in W/m^2),
-# and writes the results to a new netCDF file. This netCDF files will be used to create the 
+# and writes the results to a new netCDF file. These netCDF files will be used to create the 
 # meteogrological forcing for SOLWEIG. Note that RAIN is set to 0.
 # =============================================================================
 
@@ -632,11 +633,11 @@ def _tile_size_m(poly):
     h = _haversine_m(miny, cx, maxy, cx) 
     return w, h
 
-def process_metfiles(netcdf_file, raster_folder, base_path, selected_date_str):
+def process_metfiles(netcdf_file, raster_folder, base_path, selected_date_str, preprocess_dir):
     """
     Minimal-edits, curvilinear-safe version of your function.
     """
-    metfiles_folder = os.path.join(base_path, "metfiles")
+    metfiles_folder = os.path.join(preprocess_dir, "metfiles")
     os.makedirs(metfiles_folder, exist_ok=True)
     
     tf = TimezoneFinder()
@@ -848,16 +849,17 @@ def process_metfiles(netcdf_file, raster_folder, base_path, selected_date_str):
 # Function to process own met file: copies the source met file into new files
 # renaming each copy based on the numeric suffix extracted from .tif files.
 # =============================================================================
-def create_met_files(base_path, source_met_file):
+def create_met_files(base_path, source_met_file, preprocess_dir):
     """
     Copy a given met file to multiple outputs based on the raster tile filenames.
 
     Parameters:
-        base_path (str): Base directory containing "Building_DSM" and where to create "metfiles".
+        base_path (str): Base directory containing input rasters.
         source_met_file (str): Path to user-provided met file.
+        preprocess_dir (str): Directory for preprocessing outputs (pre_processing_outputs).
     """
-    raster_folder = os.path.join(base_path, 'Building_DSM')
-    target_folder = os.path.join(base_path, 'metfiles')
+    raster_folder = os.path.join(preprocess_dir, 'Building_DSM')
+    target_folder = os.path.join(preprocess_dir, 'metfiles')
 
     if not os.path.exists(target_folder):
         os.makedirs(target_folder)
@@ -883,12 +885,12 @@ def create_met_files(base_path, source_met_file):
[…]
