### 2026-01-16 9c944a5
Add preprocess_dir parameter to create_tiles calls
