### 2026-01-16 616dbec
Add preprocess_dir parameter to create_tiles calls
