### 2026-01-16 961ff52
Pass base_path to create_met_files function
