### 2026-01-16 24e5ce7
Update CLI example for ERA5 data on Windows
