### 2026-01-16 4c0610c
Update README for Windows CLI example and landcover file
