### 2026-01-17 7552b81
Update JOSS badge link in README.md
