+++ README.md
@@ -12,7 +12,7 @@
   <a href="https://doi.org/10.5281/zenodo.17048978"><img src="https://img.shields.io/badge/DOI-10.5281%2Fzenodo.17048978-%23ff6b6b.svg" alt="DOI"></a>
   <a href="https://www.gnu.org/licenses/gpl-3.0"><img src="https://img.shields.io/badge/License-GPLv3-%230ab5b3.svg" alt="License: GPL v3"></a>
   <a href="https://pepy.tech/project/solweig-gpu"><img src="https://img.shields.io/badge/Downloads-PyPI-%23a9b1bd.svg" alt="PyPI Downloads"></a>
-  <a href="https://joss.theoj.org/papers/27faa2bf5f6058d981df8b565f8e9a34"><img src="https://img.shields.io/badge/Published-JOSS-%23dce8f2.svg" alt="JOSS Paper"></a>
+  <a href="https://joss.theoj.org/papers/27faa2bf5f6058d981df8b565f8e9a34"><img src="https://joss.theoj.org/papers/27faa2bf5f6058d981df8b565f8e9a34/status.svg"></a>
   <a href="https://github.com/nvnsudharsan/solweig-gpu/actions/workflows/tests.yml"><img src="https://img.shields.io/badge/Tests-Passing-%23ffb703.svg" alt="Tests"></a>
 </p>
 
