### 2026-01-17 9fa7737
Update PyPI Downloads badge in README.md
