### 2026-01-17 e7f4a75
Update PyPI downloads badge color in README
