### 2026-01-17 nvnsudharsan/SOLWEIG-GPU#85 @nvnsudharsan
The base path can be set to the directory where you want the outputs saved, which may differ from the location of the input rasters. In such cases, simply provide the full path to each input raster file. The code has been refactored to automatically create a folder named processed_inputs to store the tiled input rasters. Please note the current version is 1.2.19
