### 2026-01-19 nvnsudharsan/SOLWEIG-GPU#84 @nvnsudharsan
This is intentional and consistent with commonly available land-cover datasets. Categories 3 and 4 are automatically mapped to grass because the model requires the underlying land cover beneath the canopy.
