### 2026-01-19 6de5517
Update DOI badge in README.md
