### 2026-01-19 #9535 @ruedi99ms
From my side only one issue remaining, 4 have been solved
