### 2026-01-20 4c59c2e
Update download link for sample data
update the link to match the zenodo version
