### 2026-01-20 1cab1d1
Bump version to 1.2.20
