### 2026-01-21 #9535 @nvnsudharsan
@adithirgis, @lukasbeuster, and @ruedi99ms , we have updated the codebase based on the review comments from @ruedi99ms, bumped the version to v1.2.20, and updated the sample dataset on Zenodo: https://zenodo.org/records/18283037.
