### 2026-01-23 #9535 @lukasbeuster
@nvnsudharsan Minor comment re documentation:

**Community guidelines**
Contribution guidelines are clear. If would be good to just mention how users can report issues or seek support.
