### 2026-01-23 #9535 @lukasbeuster
Hi @adithirgis

Just a quick update: I’ve started the review, and things are looking good so far. Installation and documentation seem solid. I’m still working through the functionality checks (currently downloading the test data), and will continue over the next few days.

I’ll post further comments as I go.
