### 2026-01-28 b06fee4
Update README with Windows note for SOLWEIG-GPU
Added Windows usage note for SOLWEIG-GPU regarding the main() function and multiprocessing.
