### 2026-01-28 3d0674b
Update Windows note formatting in README.md
