### 2026-01-28 f8b8a5d
Revise Windows note for SOLWEIG-GPU usage
Updated the note for Windows users to improve clarity.
