### 2026-01-29 4152c53
Update Windows user note for SOLWEIG-GPU
Clarified note for Windows users regarding multiprocessing issues with SOLWEIG-GPU.
