### 2026-01-29 b4535d4
Add Windows user note for thermal_comfort usage
Added note for Windows users regarding the use of thermal_comfort() in a main() function to avoid multiprocessing issues.
