+++ docs/quickstart.md
@@ -80,6 +80,42 @@ thermal_comfort(
 When using wrfout, the `start_time` and `end_time` should be the first and last time stamps in the wrfout dataset. The package won't compare the selected `start_time` and `end_time` to the wrfout file timestamps and automatically fetch the corresponding data. 
 `start_time` and `end_time` must be in **UTC**. The package will automatically convert to local time based on the geographic location of your study area.
 
+### Note for Windows Users: 
+If you run SOLWEIG-GPU on Windows, call thermal_comfort() inside a main() function and use if __name__ == "__main__": (see example below). This avoids multiprocessing issues (e.g., BrokenProcessPool) on Windows, where the spawn process start method re-imports the main module and can fail when multiprocessing code is executed at the top level.
+```python
+from solweig_gpu import thermal_comfort
+import multiprocessing as mp
+
+def main():
+    thermal_comfort(
+        base_path='/path/to/input',
+        selected_date_str="2020-08-13",
+        building_dsm_filename="Building_DSM.tif",
+        dem_filename="DEM.tif",
+        trees_filename="Trees.tif",
+        landcover_filename="Landcover.tif",
+        tile_size=1000,
+        overlap=100,
+        use_own_met=False,
+        own_met_file='/path/to/met.txt',  # placeholder; ignored when use_own_met=False
+        start_time="2020-08-13 06:00:00",
+        end_time="2020-08-14 05:00:00",
+        data_source_type="era5",
+        data_folder='/path/to/era5_or_wrfout',
+        save_tmrt=False,
+        save_svf=False,
+        save_kup=False,
+        save_kdown=False,
+        save_lup=False,
+        save_ldown=False,
+        save_shadow=False,
+    )
+
+if __name__ == "__main__":
+    mp.freeze_support()  # needed for Windows when using multiprocessing (esp. packaged executables)
+    main()
+```
+
 ## Command-Line Usage
 
 ```bash
