### 2026-01-29 98d78fd
Remove comment on freeze_support for clarity
