### 2026-01-29 #9535 @nvnsudharsan
> [@nvnsudharsan](https://github.com/nvnsudharsan) Hello there. Great piece of software as far as I can tell so far. I have a couple of comments re the paper:
>
> **Statement of Need** It would be great if you could more explicitly name the intended target audience and typical use cases (e.g., researchers, practitioners, city-scale thermal comfort analyses...)
>
> **State of the Field** You compare SOLWEIG-GPU and CPU-based SOLWEIG, but a broader discussion of related works is still missing. It would be great if you could briefly position it relative to other popular thermal comfort/radiation modelling tools (RayMan, ENVI-met, other UMEP components). This can be high-level, just so that readers understand what niche you occupy as well as the contribution to the field.
>
> Otherwise, the paper is looking good.

@lukasbeuster Thank you for the comments. We have updated the Summary section to explicitly identify the target audience and typical use cases, and to briefly position SOLWEIG-GPU relative to existing tools.
