### 2026-01-30 nvnsudharsan/SOLWEIG-GPU#83 @ruedi99ms
Perfect :)
