### 2026-02-03 #9535 @ruedi99ms
Everything is working fine, in my opinion

One final thing: The GUI usage is not included in the docs (only in the readme) - it would be nice to have that there as well
