### 2026-02-03 4850539
Document GDAL import error and verification steps
Added a section for common GDAL import errors and verification steps.
