### 2026-02-03 e45da12
Fix formatting in installation documentation
Remove unnecessary code block for GDAL import error section.
