### 2026-02-03 7fbcb0f
Fix formatting in installation verification section
