### 2026-02-03 #9535 @nvnsudharsan
> Everything is working fine, in my opinion
>
> One final thing: The GUI usage is not included in the docs (only in the readme) - it would be nice to have that there as well

@ruedi99ms added to the quick start guide in docs.
