### 2026-02-03 35a6198
Update installation instructions for GDAL
Added installation solutions and guidelines for GDAL in various environments.
