### 2026-02-03 cf44069
Enhance advanced workaround instructions for GDAL
Updated formatting and improved clarity of the advanced workaround section.
