### 2026-02-03 cc86cfd
Fix formatting for pip install warning in installation.md
