### 2026-02-04 #9535 @lukasbeuster
Hi all, quick validation update. Testing has been slower due to limited hardware availability. While running on Windows, I am encountering a PermissionError related to concurrent file access and am currently investigating whether this is environment-specific or a structural issue.
