### 2026-02-04 #9535 @nvnsudharsan
Hi @lukasbeuster, Thank you for the update. @ruedi99ms also had a similar issue while running on Windows, and he suggested calling thermal_comfort() inside a main() function (https://github.com/nvnsudharsan/SOLWEIG-GPU/issues/83).
