### 2026-02-05 #9535 @lukasbeuster
> Hi [@lukasbeuster](https://github.com/lukasbeuster), Thank you for the update. [@ruedi99ms](https://github.com/ruedi99ms) also had a similar issue while running on Windows, and he suggested calling thermal_comfort() inside a main() function ([nvnsudharsan/SOLWEIG-GPU#83](https://github.com/nvnsudharsan/SOLWEIG-GPU/issues/83)).

That resolved the problem on my end as well. Thank you for pointing this out. This seems to be a classic Windows multiprocessing issue: Python uses the spawn start method, which re-imports the script in each child process. Without proper guarding, this causes thermal_comfort() to be executed repeatedly. I see this is now reflected in the documentation; a brief explanation of the underlying reason might be helpful for users less familiar with multiprocessing flows on Windows (both ruedi99ms and I stumbled across this...).

With that resolved, testing was successful. Overall, this is a solid and well-executed contribution, and I'm looking forward to using it in my own research.

This concludes my review @adithirgis. Thank you for your patience with my somewhat slow turnaround.

If I may suggest a few items for the backlog:

- 	The ability to limit or manage shared resource usage (CPU cores, VRAM), particularly when running in remote or shared environments. This could be a dedicated input or clearly documented with examples outside of the main functions.
- 	The option to define a custom output directory, rather than relying on using the Input_rasters directory.
- 	Potential I/O bottlenecks when writing large intermediate outputs. I am not sure to what extent this can be addressed or streamlined, but it may be worth investigating in future releases.

I would also recommend a final pass over the documentation to remove duplication and outdated or incomplete references (for example, the sample dataset link [here](https://solweig-gpu.readthedocs.io/en/latest/testing.html#test-with-sample-data).

And last but not least, while I understand the sizing constraints imposed by the forcing datasets, a smaller test dataset would have been very helpful for quick validation and onboarding.
