### 2026-02-06 #9535 @adithirgis
Thanks @lukasbeuster and @ruedi99ms for the review!
