### 2026-02-09 5c58a7f
Enhance documentation with issue reporting and support guidelines
Added a section on reporting issues and getting support in the developer guide and contributing guide. Updated output file structure references to reflect changes in naming conventions and directory organization. Improved clarity in the testing guide regarding verification and test suite execution. Adjusted quickstart instructions to align with new output folder structure.
