+++ docs/outputs.md
@@ -4,26 +4,26 @@ This guide explains the output files generated by SOLWEIG-GPU and how to interpr
 
 ## Output Structure
 
-After running a simulation, outputs are saved in the `Outputs/` directory within your `base_path`. The structure is organized by tiles:
+After running a simulation, outputs are saved in the `output_folder/` directory within your `base_path`. The structure is organized by tile key (origin coordinates in the raster grid):
 
 ```
 base_path/
-└── Outputs/
-    ├── tile_0_0/
+└── output_folder/
+    ├── 0_0/
     │   ├── UTCI_0_0.tif
-    │   ├── Tmrt_0_0.tif
+    │   ├── TMRT_0_0.tif
     │   └── SVF_0_0.tif
-    ├── tile_0_1000/
+    ├── 0_1000/
     │   ├── UTCI_0_1000.tif
-    │   ├── Tmrt_0_1000.tif
+    │   ├── TMRT_0_1000.tif
     │   └── SVF_0_1000.tif
-    └── tile_1000_0/
+    └── 1000_0/
         ├── UTCI_1000_0.tif
-        ├── Tmrt_1000_0.tif
+        ├── TMRT_1000_0.tif
         └── SVF_1000_0.tif
 ```
 
-Each tile is named according to its origin coordinates in the raster grid (e.g., `tile_0_0`, `tile_0_1000`).
+Each subfolder is named by the tile key (e.g., `0_0`, `0_1000`, `1000_0`). File names use `TMRT` (not Tmrt) to match the written rasters.
 
 ## Output Files
 
@@ -54,7 +54,7 @@ UTCI is always computed and saved. It provides a comprehensive measure of therma
 
 ### Mean Radiant Temperature (Tmrt)
 
-**Filename:** `Tmrt_X_Y.tif`  
+**Filename:** `TMRT_X_Y.tif`  
 **Format:** Multi-band GeoTIFF  
 **Units:** °C  
 **Description:** The mean radiant temperature in a standing position.
@@ -65,14 +65,14 @@ Tmrt represents the uniform temperature of an imaginary enclosure in which the r
 
 ### Sky View Factor (SVF)
 
-**Filename:** `SVF_X_Y.tif`  
+**Filename:** `SVF_X_Y.tif` (single-band; X_Y is the tile key)  
 **Format:** Single-band GeoTIFF  
 **Units:** Dimensionless (0-1)  
 **Description:** The fraction of the sky hemisphere visible from each point.
 
 SVF is a geometric parameter that quantifies the openness of a location to the sky. It affects both shortwave and longwave radiation exchange.
 
-**Saved when:** `save_tmrt=True`
+**Saved when:** `save_svf=True`
 
 !!! Note:
     Unlike other outputs, SVF is a geometric property that doesn't change with time. It is saved as a single-band raster, not multi-band.
@@ -157,7 +157,7 @@ import matplotlib.pyplot as plt
 import numpy as np
 
 # Open the UTCI file
-ds = gdal.Open('Outputs/tile_0_0/UTCI_0_0.tif')
+ds = gdal.Open('output_folder/0_0/UTCI_0_0.tif')
 
 # Read a specific hour (e.g., hour 12 = noon)
 band = ds.GetRasterBand(12)
@@ -195,7 +195,7 @@ import numpy as np
 from tempfile import TemporaryDirectory
 from osgeo import gdal
 
-base_dir = r"path_to_outputs/Outputs" # Locate the simulation output
+base_dir = r"path_to_base/output_folder"  # Locate the simulation output (base_path/output_folder)
[…]
