### 2026-02-09 58c9a4f
Update contributing and installation documentation to include test setup instructions
Revised the contributing guide to add steps for setting up the development environment and running the test suite. Updated the installation guide to clarify the installation of test dependencies and how to verify the installation by running tests.
