### 2026-02-09 425e37b
Update documentation for output directory naming and Windows user instructions
Changed output directory reference from `Outputs/` to `output_folder/` in README and examples. Revised Windows user guidance to clarify multiprocessing issues and the need for a main function when calling `thermal_comfort()`. Updated example notebook references and sample data download link in examples README.
