### 2026-02-09 f9d63a6
Update output directory references in code and testing documentation
Changed all instances of `Outputs/` to `output_folder/` in the `thermal_comfort` function documentation and the testing setup. This aligns the code with the updated output structure and ensures consistency across documentation.
