### 2026-02-09 70f94fa
Enhance documentation for output directory configuration
Updated the documentation across multiple files to clarify how to set the `base_path` for output files. Added tips for writing outputs to a different directory and emphasized the need for complete paths for input rasters. This improves user guidance on configuring output locations and aligns with recent changes in the output structure.
