+++ docs/outputs.md
@@ -4,7 +4,7 @@ This guide explains the output files generated by SOLWEIG-GPU and how to interpr
 
 ## Output Structure
 
-After running a simulation, outputs are saved in the `output_folder/` directory within your `base_path`. The structure is organized by tile key (origin coordinates in the raster grid):
+After running a simulation, outputs are saved in the `output_folder/` directory within your `base_path`. To write outputs elsewhere, set `base_path` to the desired directory and pass complete paths for the input rasters (Building DSM, DEM, Trees, land cover); see [Configuration](configuration.md#base_path). The structure is organized by tile key (origin coordinates in the raster grid):
 
 ```
 base_path/
