### 2026-02-09 1927f44
Refine documentation for `base_path` and raster filenames
Updated the `thermal_comfort` function documentation to clarify the usage of `base_path` for output directories and relative raster paths. Enhanced descriptions for raster filename parameters to indicate they can be specified as complete paths. This improves user understanding of file management within the application.
