### 2026-02-09 33898a4
Enhance documentation clarity and consistency
Updated various documentation files to improve clarity and consistency. Changes include refining descriptions for file creation in tests, correcting formatting and phrasing in installation and configuration guides, and enhancing the presentation of output data types. These updates aim to provide clearer guidance for users and contributors.
