### 2026-02-09 e810917
Add preprocessing and UTCI processing functions
Introduced new functions: `preprocess`, `run_walls_aspect`, and `run_utci_tiles` to enhance the workflow for raster data processing. The `preprocess` function validates rasters and prepares metfiles, while `run_walls_aspect` calculates wall height and aspect for tiles. The `run_utci_tiles` function executes UTCI calculations for specified tiles, allowing for flexible output options. Updated the `__init__.py` to include these new functions for easier access.
