+++ solweig_gpu/solweig_gpu.py
@@ -10,7 +10,189 @@
 #but WITHOUT ANY WARRANTY; without even the implied warranty of
 #MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
 #GNU General Public License for more details.
-from typing import Optional
+from typing import Optional, List
+
+# Lazy imports used below to avoid loading heavy deps at import time.
+
+
+def preprocess(
+    base_path: str,
+    selected_date_str: str,
+    building_dsm_filename: str = 'Building_DSM.tif',
+    dem_filename: str = 'DEM.tif',
+    trees_filename: str = 'Trees.tif',
+    landcover_filename: Optional[str] = None,
+    tile_size: int = 3600,
+    overlap: int = 20,
+    use_own_met: bool = True,
+    start_time: Optional[str] = None,
+    end_time: Optional[str] = None,
+    data_source_type: Optional[str] = None,
+    data_folder: Optional[str] = None,
+    own_met_file: Optional[str] = None,
+    preprocess_dir: Optional[str] = None,
+) -> str:
+    """
+    Run preprocessing only: validate rasters, create tiles, and prepare metfiles.
+
+    Use this when you want to run preprocessing once and then call
+    :func:`run_walls_aspect` and :func:`run_utci_tiles` separately (e.g. for
+    multiple dates or a subset of tiles).
+
+    Args:
+        base_path: Base directory; used to resolve relative raster paths.
+        selected_date_str: Simulation date 'YYYY-MM-DD'.
+        building_dsm_filename, dem_filename, trees_filename, landcover_filename:
+            Raster paths or filenames (relative to base_path or absolute).
+        tile_size: Tile size in pixels.
+        overlap: Overlap between tiles in pixels.
+        use_own_met: If True, use own_met_file; else use ERA5/WRF (requires
+            start_time, end_time, data_source_type, data_folder).
+        start_time, end_time: Required for ERA5/WRF (UTC 'YYYY-MM-DD HH:MM:SS').
+        data_source_type: 'ERA5' or 'wrfout' when use_own_met is False.
+        data_folder: Folder with ERA5/WRF NetCDF files when use_own_met is False.
+        own_met_file: Path to custom met file when use_own_met is True.
+        preprocess_dir: Directory for preprocessing outputs. Defaults to
+            ``{base_path}/processed_inputs``.
+
+    Returns:
+        The path to the preprocessing directory (tiles and metfiles written there).
+    """
+    import os
+    from .preprocessor import ppr
+
+    if preprocess_dir is None:
+        preprocess_dir = os.path.join(base_path, "processed_inputs")
+    os.makedirs(preprocess_dir, exist_ok=True)
+
+    ppr(
+        base_path, building_dsm_filename, dem_filename, trees_filename,
+        landcover_filename, tile_size, overlap, selected_date_str, use_own_met,
+        start_time, end_time, data_source_type, data_folder, own_met_file,
+        preprocess_dir=preprocess_dir,
+    )
+    return preprocess_dir
+
+
+def run_walls_aspect(preprocess_dir: str) -> None:
+    """
+    Run wall height and aspect calculation for all tiles in the preprocessing directory.
+
+    Call this after :func:`preprocess`. Writes to ``{preprocess_dir}/walls`` and
+    ``{preprocess_dir}/aspect``.
+
+    Args:
+        preprocess_dir: Path returned by :func:`preprocess` (contains
+            Building_DSM/, DEM/, Trees/, etc.).
+    """
+    import os
+    from .walls_aspect import run_parallel_processing
+
[…]
