### 2026-02-09 ee2bf82
Clarify language in input_data.md
