### 2026-02-09 3449269
Update wording for clarity in examples section
