### 2026-02-09 #9535 @nvnsudharsan
> > Hi [@lukasbeuster](https://github.com/lukasbeuster), Thank you for the update. [@ruedi99ms](https://github.com/ruedi99ms) also had a similar issue while running on Windows, and he suggested calling thermal_comfort() inside a main() function ([nvnsudharsan/SOLWEIG-GPU#83](https://github.com/nvnsudharsan/SOLWEIG-GPU/issues/83)).
>
> That resolved the problem on my end as well. Thank you for pointing this out. This seems to be a classic Windows multiprocessing issue: Python uses the spawn start method, which re-imports the script in each child process. Without proper guarding, this causes thermal_comfort() to be executed repeatedly. I see this is now reflected in the documentation; a brief explanation of the underlying reason might be helpful for users less familiar with multiprocessing flows on Windows (both ruedi99ms and I stumbled across this...).
>
> With that resolved, testing was successful. Overall, this is a solid and well-executed contribution, and I'm looking forward to using it in my own research.
>
> This concludes my review [@adithirgis](https://github.com/adithirgis). Thank you for your patience with my somewhat slow turnaround.
>
> If I may suggest a few items for the backlog:
>
> * The ability to limit or manage shared resource usage (CPU cores, VRAM), particularly when running in remote or shared environments. This could be a dedicated input or clearly documented with examples outside of the main functions.
> * The option to define a custom output directory, rather than relying on using the Input_rasters directory.
> * Potential I/O bottlenecks when writing large intermediate outputs. I am not sure to what extent this can be addressed or streamlined, but it may be worth investigating in future releases.
>
> I would also recommend a final pass over the documentation to remove duplication and outdated or incomplete references (for example, the sample dataset link [here](https://solweig-gpu.readthedocs.io/en/latest/testing.html#test-with-sample-data).
>
> And last but not least, while I understand the sizing constraints imposed by the forcing datasets, a smaller test dataset would have been very helpful for quick validation and onboarding.

Thank you very much for the thorough and thoughtful review, and for taking the time to test the workflow so carefully. We really appreciate your detailed feedback and are glad to hear that you plan to use the tool in your own research.

We have addressed several of your suggestions in the latest updates:

- We have expanded the documentation with a clearer explanation of the Windows “spawn” behavior and the need for proper guarding, to make this more accessible for users less familiar with multiprocessing internals.
- We have now uploaded  ERA5 and ownmet forcing datasets separately on Zenodo. The large size of the original sample is primarily due to the high spatial and temporal resolution of the WRF forcing, but the ERA5/ownmet options should make quick validation and onboarding much easier.
- The documentation has been updated to clarify that users can set the base directory as an output directory. In this case, the input datasets should be provided with full paths. We agree that this flexibility is important, especially in shared or remote environments.
- While explicit CPU/VRAM limiting is not yet exposed as a dedicated interface, this is now on our development roadmap. In a future release, we plan to allow users to call individual processing functions independently, so that only the required components are executed.
- We have done a final pass through the documentation to remove duplication and to update or complete any outdated references, including dataset links.

Thank you again for the constructive review and for your patience during the process.

Both of your feedback has genuinely helped improve the usability and clarity of the package.
