### 2026-02-09 #9535 @nvnsudharsan
> 👋 [@nvnsudharsan](https://github.com/nvnsudharsan) - we are almost there! Next is just setting up the archive for your new release.
>
> We want to make sure the archival has the correct metadata that JOSS requires. This includes a title that matches the paper title and a correct author list.
>
> So here is what we have left to do:
>
> * [ ]
>   [ ]   Confirm if the issues raised by the reviewers have been resolved.
>   [ ]
>   [ ]   Confirm that the title of your paper in your paper.md file has no special formatting (e.g., backticks) as this will cause formatting issues when accepted.
>   [ ]
>   [ ]   Conduct a GitHub release of the current reviewed version of the software and archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository).  Please ensure that the software archive uses the same license as the license you have posted in your code repository.
>   [ ]
>   [ ]   Check the archival deposit (e.g., in Zenodo) to ensure it has the correct metadata. This includes the title (should match the paper title exactly including the case) and the author list (should match the people and the order in which they are listed on your paper exactly).
>   [ ]
>   [ ]   Please respond with the DOI of the archived version here along with the version tag name released.
>
> I can then move forward with recommending for accepting the submission.

Hi @adithirgis,

All the listed items have been completed:
- The reviewer comments have been fully addressed.
- The paper title in paper.md has been checked and contains no special formatting.
- A GitHub release has been created for the reviewed version, and the software has been archived on Zenodo using the same license as the repository.
- The Zenodo record has been verified to ensure the title and author list exactly match the paper (including order and case).
- The Zenodo DOI for the archived release is: https://doi.org/10.5281/zenodo.18561860
- The corresponding version tag is: v1.2.21

Please let me know if anything else is needed. Thanks.
