### 2026-02-11 #9535 @nvnsudharsan
> Hi @nvnsudharsan! The zenodo version number is correct - it matches the released version. But there is no code on Zenodo. Can you please upload the same version of code on Zenodo? Also the affiliations should be exactly same in Zenodo. Please update both. Thanks

Hi @adithirgis The code is included in the Zenodo record as SOLWEIG-GPU_1_2_21.zip, along with the sample data. We will update the affiliations to match exactly, and may create a new version to keep everything consistent. Thanks!
