### 2026-02-12 #9535 @nvnsudharsan
Hi @adithirgis
Harsh G Kamath: Refactoring GPU code (co-lead), Packaging (co-lead), Documentation (co-lead)
Naveen Sudharsan: Refactoring GPU code (co-lead), Packaging (co-lead), Documentation (co-lead)
Manmeet Singh: Refactoring GPU code, Packaging (co-lead)
Nils Wallenberg: SOLWEIG development, testing, and code
Fridrick Lindberg: SOLWEIG development, testing, and code
Dev Niyogi: Model Architecture
