### 2026-02-13 #9535 @editorialbot
:wave: @openjournals/ese-eics, this paper is ready to be accepted and published.
